refactor(dbhelper): split monolithic 4406-line file into organized modules
Split the massive utils/dbhelper/dbhelper.go into 8 focused files organized
by functional responsibility:

- types.go (906 lines): All struct definitions and methods
- replication.go (1159 lines): Replication control operations
- status.go (430 lines): Server status and variable queries
- performance.go (363 lines): Performance Schema and query analysis
- binlog.go (314 lines): Binary log operations
- transaction.go (129 lines): Transaction control and locking
- connection.go (97 lines): Connection management
- dbhelper.go (31 lines): Package documentation

Benefits:
- Improved code organization and maintainability
- Better testability with focused modules
- Easier navigation and code discovery
- Clear separation of concerns
- Reduced cognitive load for developers

Original file preserved as reference (not committed).
All 162 exported functions categorized and extracted.
Zero compilation errors, full backward compatibility maintained.
