+++ utils/dbhelper/dbhelper.go
@@ -7,4400 +7,25 @@
 // an additional term, ALL code must carry the original Author(s) credit in comment form.
 // See LICENSE in this directory for the integral text.
 
-package dbhelper
-
-import (
-	"context"
-	"database/sql"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"hash/crc64"
-	"log"
-	"net"
-	"regexp"
-	"slices"
-	"sort"
-	"strconv"
-	"strings"
-	"sync"
-	"sync/atomic"
-	"time"
-
-	"github.com/jmoiron/sqlx"
-	"github.com/percona/go-mysql/query"
-	"github.com/signal18/replication-manager/utils/misc"
-	"github.com/signal18/replication-manager/utils/version"
-)
-
-const debug = false
-const (
-	DDMMYYYYhhmmss = "2006-01-02 15:04:05"
-)
-
-type Plugin struct {
-	Name    string         `json:"name"`
-	Status  string         `json:"status"`
-	Type    string         `json:"type"`
-	Library sql.NullString `json:"library"`
-	License string         `json:"license"`
-}
-
-type chunk struct {
-	ChunkId       uint64 `json:"chunkId"`
-	ChunkMinKey   string `json:"chunkMinKey"`
-	ChunkMaxKey   string `json:"chunkMaxKey"`
-	ChunkCheckSum uint64 `json:"chunkCheckSum"`
-}
-
-type MetaDataLock struct {
-	Thread_id     uint64         `json:"threadId" db:"THREAD_ID"`
-	Lock_mode     sql.NullString `json:"lockMode" db:"LOCK_MODE"`
-	Lock_duration sql.NullString `json:"lockDuration" db:"LOCK_DURATION"`
-	Lock_time_ms  sql.NullInt64  `json:"lockTimeMs" db:"LOCK_TIME_MS"`
-	Lock_type     sql.NullString `json:"lockType" db:"LOCK_TYPE"`
-	Lock_catalog  sql.NullString `json:"lockCatalog" db:"TABLE_CATALOG"`
-	Lock_schema   sql.NullString `json:"lockSchema" db:"TABLE_SCHEMA"`
-	Lock_name     sql.NullString `json:"lockName" db:"TABLE_NAME"`
-}
-
-type ResponseTime struct {
-	Time  string `json:"time" db:"TIME"`
-	Count uint64 `json:"count" db:"COUNT"`
-	Total string `json:"total" db:"TOTAL"`
-}
-
-type PFSQuery struct {
-	Digest           string          `json:"digest"`
-	Query            string          `json:"query"`
-	Digest_text      string          `json:"digestText"`
-	Schema_name      string          `json:"shemaName"`
-	Last_seen        string          `json:"lastSeen"`
-	Plan_full_scan   string          `json:"planFullScan"`
-	Plan_tmp_disk    int64           `json:"planTmpDisk"`
-	Plan_tmp_mem     int64           `json:"planTmpMem"`
-	Exec_count       int64           `json:"execCount"`
-	Err_count        int64           `json:"errCount"`
-	Warn_count       int64           `json:"warnCount"`
-	Exec_time_total  string          `json:"execTimeTotal"`
-	Exec_time_max    sql.NullFloat64 `json:"execTimeMax"`
-	Exec_time_avg_ms sql.NullFloat64 `json:"execTimeAvgMs"`
-	Rows_sent        int64           `json:"rowsSent"`
-	Rows_sent_avg    int64           `json:"rowsSentAvg"`
-	Rows_scanned     int64           `json:"rowsScanned"`
-	Value            string          `json:"value"`
-}
-
-type PFSQuerySorter []PFSQuery
-
-func (a PFSQuerySorter) Len() int      { return len(a) }
-func (a PFSQuerySorter) Swap(i, j int) { a[i], a[j] = a[j], a[i] }
-func (a PFSQuerySorter) Less(i, j int) bool {
-	l, _ := strconv.ParseFloat(a[i].Value, 64)
-	r, _ := strconv.ParseFloat(a[j].Value, 64)
-	return l > r
-}
-
-type Column struct {
-	Name      string  `json:"name"`
-	Type      string  `json:"type"`
-	Nullable  bool    `json:"nullable"`
-	Default   *string `json:"default,omitempty"`
-	Extra     string  `json:"extra,omitempty"`
-	Charset   *string `json:"charset,omitempty"`
-	Collation *string `json:"collation,omitempty"`
-	Crc64     uint64  `json:"crc64,omitempty"`
-}
-
-type IndexColumn struct {
-	Name   string  `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
-	Prefix *uint16 `protobuf:"varint,2,opt,name=prefix,proto3" json:"prefix,omitempty"`
-}
-
-type Index struct {
-	Name    string        `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
-	Unique  bool          `protobuf:"varint,2,opt,name=unique,proto3" json:"unique,omitempty"`
-	Type    string        `protobuf:"bytes,3,opt,name=type,proto3" json:"type,omitempty"`
-	Crc64   uint64        `protobuf:"varint,4,opt,name=crc64,proto3" json:"crc64,omitempty"`
-	Columns []IndexColumn `protobuf:"bytes,4,opt,name=columns,proto3" json:"columns,omitempty"`
-}
-
-type Table struct {
-	TableSchema       string             `protobuf:"bytes,1,opt,name=table_schema,json=tableSchema,proto3" json:"table_schema,omitempty"`
-	TableName         string             `protobuf:"bytes,2,opt,name=table_name,json=tableName,proto3" json:"table_name,omitempty"`
-	Engine            string             `protobuf:"bytes,3,opt,name=engine,proto3" json:"engine,omitempty"`
-	TableRows         int64              `protobuf:"varint,4,opt,name=table_rows,json=tableRows,proto3" json:"table_rows,omitempty"`
-	DataLength        int64              `protobuf:"varint,5,opt,name=data_length,json=dataLength,proto3" json:"data_length,omitempty"`
-	IndexLength       int64              `protobuf:"varint,6,opt,name=index_length,json=indexLength,proto3" json:"index_length,omitempty"`
-	TableCrc          uint64             `protobuf:"varint,7,opt,name=table_crc,json=tableCrc,proto3" json:"table_crc,omitempty"`
-	TableClusters     string             `protobuf:"bytes,8,opt,name=table_clusters,json=tableClusters,proto3" json:"table_clusters,omitempty"`
-	TableSync         string             `protobuf:"bytes,9,opt,name=table_sync,json=tableSync,proto3" json:"table_sync,omitempty"`
-	TableColumns      []Column           `protobuf:"bytes,10,opt,name=table_columns,json=tableColumns,proto3" json:"table_columns,omitempty"`
-	TableIndexes      []Index            `protobuf:"bytes,11,opt,name=table_indexes,json=tableIndexes,proto3" json:"table_indexes,omitempty"`
-	TableColumnsCrc64 uint64             `protobuf:"varint,12,opt,name=table_columns_crc64,json=tableColumnsCrc64,proto3" json:"table_columns_crc64,omitempty"`
-	TableIndexesCrc64 uint64             `protobuf:"varint,13,opt,name=table_indexes_crc64,json=tableIndexesCrc64,proto3" json:"table_indexes_crc64,omitempty"`
-	TableColumnMap    map[string]*Column `protobuf:"-" json:"-"`
-	TableIndexMap     map[string]*Index  `protobuf:"-" json:"-"`
-}
-
-func (x *Table) CanonicalizeColumns() {
-	// columns
-	sort.Slice(x.TableColumns, func(i, j int) bool {
-		return x.TableColumns[i].Name < x.TableColumns[j].Name
-	})
-}
-
-func (x *Table) CanonicalizeIndexes() {
-	// indexes
-	sort.Slice(x.TableIndexes, func(i, j int) bool {
-		return x.TableIndexes[i].Name < x.TableIndexes[j].Name
-	})
-
-	// index columns (order matters!)
-	for i := range x.TableIndexes {
-		sort.Slice(x.TableIndexes[i].Columns, func(a, b int) bool {
-			return x.TableIndexes[i].Columns[a].Name <
-				x.TableIndexes[i].Columns[b].Name
-		})
-	}
-}
-
-func (x *Table) BuildColumnMap() {
-	x.CanonicalizeColumns()
-	// columns
-	colMap := make(map[string]*Column, len(x.TableColumns))
-	for _, c := range x.TableColumns {
-		colMap[c.Name] = &c
-	}
-	x.TableColumnMap = colMap
-}
-
-func (x *Table) BuildIndexMap() {
-	x.CanonicalizeIndexes()
-	// indexes
-	idxMap := make(map[string]*Index, len(x.TableIndexes))
-	for _, i := range x.TableIndexes {
-		idxMap[i.Name] = &i
-	}
-	x.TableIndexMap = idxMap
-}
-
-func (x *Table) HashColumns(crc64Table *crc64.Table) {
-	x.BuildColumnMap()
-	var tableData strings.Builder
-	for _, col := range x.TableColumnMap {
-		if tableData.Len() > 0 {
-			tableData.WriteString("||")
-		}
-
-		var columnData strings.Builder
-		columnData.WriteString(col.Name)
-		columnData.WriteString("|" + col.Type)
-		if col.Nullable {
-			columnData.WriteString("|NULLABLE")
-		} else {
-			columnData.WriteString("|")
-		}
-		if col.Default != nil {
-			columnData.WriteString("|DEFAULT " + *col.Default)
-		} else {
-			columnData.WriteString("|")
-		}
-		if col.Extra != "" {
-			columnData.WriteString("|" + col.Extra)
-		}
-		if col.Charset != nil {
-			columnData.WriteString("|CHARSET|" + *col.Charset)
-		}
-		if col.Collation != nil {
-			columnData.WriteString("|COLLATION|" + *col.Collation)
-		}
-		col.Crc64 = crc64.Checksum([]byte(columnData.String()), crc64Table)
-		tableData.WriteString(fmt.Sprintf("%d", col.Crc64))
-	}
-
-	x.TableColumnsCrc64 = crc64.Checksum([]byte(tableData.String()), crc64Table)
-}
-
-func (x *Table) HashIndexes(crc64Table *crc64.Table) {
-	x.BuildIndexMap()
-	var tableData strings.Builder
-
-	for _, idx := range x.TableIndexMap {
-		if tableData.Len() > 0 {
-			tableData.WriteString("||")
-		}
-
-		var indexData strings.Builder
-		indexData.WriteString(idx.Name)
-		if idx.Unique {
-			indexData.WriteString("|UNIQUE|")
-		} else {
-			indexData.WriteString("|NONUNIQUE|")
-		}
-		indexData.WriteString(idx.Type)
-		for _, col := range idx.Columns {
-			indexData.WriteString("|" + col.Name)
-			if col.Prefix != nil {
-				indexData.WriteString("(" + strconv.Itoa(int(*col.Prefix)) + ")")
-			}
-		}
-		indexData.WriteString("||")
-
-		idx.Crc64 = crc64.Checksum([]byte(indexData.String()), crc64Table)
-		tableData.WriteString(fmt.Sprintf("%d", idx.Crc64))
-	}
-}
-
-func (x *Table) ColumnDiffs(other *Table, replicaID string) []string {
-	var diffs []string
-	for colName, col := range x.TableColumnMap {
-		otherCol, ok := other.TableColumnMap[colName]
-		if !ok {
-			diffs = append(diffs, fmt.Sprintf("%s: missing on replica %s", colName, replicaID))
-			continue
-		}
-
-		if col.Crc64 != otherCol.Crc64 {
-			diffs = append(diffs, fmt.Sprintf("%s: different", colName))
-		}
-	}
-
-	for _, colName := range other.TableColumns {
-		if _, ok := x.TableColumnMap[colName.Name]; !ok {
-			diffs = append(diffs, fmt.Sprintf("%s: missing on master", colName.Name))
-		}
-	}
-
-	return diffs
-}
-
-func (x *Table) IndexDiffs(other *Table, replicaID string) []string {
-	var diffs []string
-	for idxName, idx := range x.TableIndexMap {
-		otherIdx, ok := other.TableIndexMap[idxName]
-		if !ok {
-			diffs = append(diffs, fmt.Sprintf("index %s: missing on replica %s", idxName, replicaID))
-			continue
-		}
-		if idx.Crc64 != otherIdx.Crc64 {
-			diffs = append(diffs, fmt.Sprintf("index %s: different with replica %s", idxName, replicaID))
-		}
-	}
-	for _, idxName := range other.TableIndexes {
-		if _, ok := x.TableIndexMap[idxName.Name]; !ok {
-			diffs = append(diffs, fmt.Sprintf("index %s: missing on master", idxName.Name))
-		}
-	}
-	return diffs
-}
-
-func (x *Table) GetTableSchema() string {
-	if x != nil {
-		return x.TableSchema
-	}
-	return ""
-}
-
-func (x *Table) GetTableName() string {
-	if x != nil {
-		return x.TableName
-	}
-	return ""
-}
-
-func (x *Table) GetEngine() string {
-	if x != nil {
-		return x.Engine
-	}
-	return ""
-}
-
-func (x *Table) GetTableRows() int64 {
-	if x != nil {
-		return x.TableRows
-	}
-	return 0
-}
-
-func (x *Table) GetDataLength() int64 {
-	if x != nil {
-		return x.DataLength
-	}
-	return 0
-}
-
-func (x *Table) GetIndexLength() int64 {
-	if x != nil {
-		return x.IndexLength
-	}
-	return 0
-}
-
-func (x *Table) GetTableCrc() uint64 {
-	if x != nil {
-		return x.TableCrc
-	}
-	return 0
-}
-
-func (x *Table) GetTableClusters() string {
-	if x != nil {
-		return x.TableClusters
-	}
-	return ""
-}
-
-func (x *Table) GetTableSync() string {
-	if x != nil {
-		return x.TableSync
-	}
-	return ""
-}
-
-type TableSizeSorter []*Table
-
-func (a TableSizeSorter) Len() int      { return len(a) }
-func (a TableSizeSorter) Swap(i, j int) { a[i], a[j] = a[j], a[i] }
-func (a TableSizeSorter) Less(i, j int) bool {
-	return a[i].DataLength+a[i].IndexLength > a[j].DataLength+a[j].IndexLength
-}
-
-type Disk struct {
-	Disk      string
-	Path      string
-	Total     int32
-	Used      int32
-	Available int32
-}
-
-/* replaced by Table
-type Table struct {
-	Table_schema   string `json:"tableSchema"`
-	Table_name     string `json:"tableName"`
-	Engine         string `json:"engine"`
-	Table_rows     int64  `json:"tableRows"`
-	Data_length    int64  `json:"dataLength"`
-	Index_length   int64  `json:"indexLength"`
-	Table_crc      uint64 `json:"tableCrc"`
-	Table_clusters string `json:"tableClusters"`
-	Table_sync     string `json:"tableSync"`
-}
-*/
-
-type Grant struct {
-	User     string `json:"user"`
-	Host     string `json:"host"`
-	Password string `json:"-"`
-	Hash     uint64 `json:"hash"`
-}
-
-type Event struct {
-	Db      string `json:"db"`
-	Name    string `json:"name"`
-	Definer string `json:"definer"`
-	Status  int64  `json:"status"`
-}
-
-type Processlist struct {
-	Id                 uint64          `json:"id" db:"Id"`
-	User               string          `json:"user" db:"User"`
-	Host               string          `json:"host" db:"Host"`
-	Db                 sql.NullString  `json:"db" db:"db"`
-	Command            string          `json:"command" db:"Command"`
-	Time               sql.NullFloat64 `json:"time" db:"Time"`
-	TimeMs             sql.NullFloat64 `json:"timeMs" db:"Time_ms"`
-	State              sql.NullString  `json:"state" db:"State"`
-	Info               sql.NullString  `json:"info" db:"Info"`
-	Progress           sql.NullFloat64 `json:"progress" db:"Progress"`
-	RowsSent           uint64          `json:"rowsSent" db:"Rows_sent"`
-	RowsExamined       uint64          `json:"rowsExamined" db:"Rows_examined"`
-	Url                string          `json:"url" db:"Url"`
-	TrxIsolation       sql.NullString  `json:"trxIsolationLevel" db:"trx_isolation_level"`
-	TrxTime            uint64          `json:"trxTime" db:"trx_time"`
-	TrxTablesInUse     uint64          `json:"txrTablesInUse" db:"trx_tables_in_use"`
-	TrxTablesLocked    uint64          `json:"trxTablesLocked" db:"trx_tables_locked"`
-	TrxLockStructs     uint64          `json:"trxLockStructs"  db:"trx_lock_structs"`
-	TrxLockMemoryBytes uint64          `json:"trxLockMemoryBytes"  db:"trx_lock_memory_bytes"`
-	TrxRowsLocked      uint64          `json:"trxRowsLocked"   db:"trx_rows_locked"`
-	TrxRowsModified    uint64          `json:"trxRowsModified"  db:"trx_rows_modified"`
-	TrxIsReadOnly      int             `json:"trxIsReadOnly"  db:"trx_is_read_only"`
-}
-
-type LogSlow struct {
-	Start_time     int64          `db:"start_time"`
-	User_host      sql.NullString `db:"user_host"`
-	Query_time     string         `db:"query_time"`
-	Lock_time      string         `db:"lock_time"`
-	Rows_sent      int            `db:"rows_sent"`
-	Rows_examined  int            `db:"rows_examined"`
-	Db             sql.NullString `db:"db"`
-	Last_insert_id int            `db:"last_insert_id"`
-	Insert_id      int            `db:"insert_id"`
-	Server_id      int            `db:"server_id"`
-	Sql_text       sql.NullString `db:"sql_text"`
-	Thread_id      int64          `db:"thread_id"`
-	Rows_affected  int            `db:"rows_affected"`
-	Digest         string
-}
-
-type SlaveHosts struct {
-	Server_id    uint64 `json:"serverId"`
-	Host         string `json:"host"`
-	Port         uint   `json:"port"`
-	Master_id    uint64 `json:"masterId"`
-	Source_id    uint64 `json:"sourceId"`
-	Replica_UUID string `json:"replicaUUID"`
-}
-
-type MasterStatus struct {
-	File              string `json:"file"`
-	Position          uint   `json:"position"`
-	Binlog_Do_DB      string `json:"binlogDoDB"`
-	Binlog_Ignore_DB  string `json:"binlogIgnoreDB"`
-	Executed_Gtid_Set string `json:"executedGtidSet"`
-}
-
-type SlaveStatus struct {
-	ConnectionName           sql.NullString `db:"Connection_name" json:"connectionName"`
-	ChannelName              sql.NullString `db:"Channel_Name" json:"channelName"`
-	MasterHost               sql.NullString `db:"Master_Host" json:"masterHost"`
-	MasterUser               sql.NullString `db:"Master_User" json:"masterUser"`
-	MasterPort               sql.NullString `db:"Master_Port" json:"masterPort"`
-	MasterLogFile            sql.NullString `db:"Master_Log_File" json:"masterLogFile"`
-	ReadMasterLogPos         sql.NullString `db:"Read_Master_Log_Pos" json:"readMasterLogPos"`
-	RelayMasterLogFile       sql.NullString `db:"Relay_Master_Log_File" json:"relayMasterLogFile"`
-	SlaveIORunning           sql.NullString `db:"Slave_IO_Running" json:"slaveIoRunning"`
-	SlaveSQLRunning          sql.NullString `db:"Slave_SQL_Running" json:"slaveSqlRunning"`
-	ExecMasterLogPos         sql.NullString `db:"Exec_Master_Log_Pos" json:"execMasterLogPos"`
-	SecondsBehindMaster      sql.NullInt64  `db:"Seconds_Behind_Master" json:"secondsBehindMaster"`
-	MasterLastEventTime      sql.NullString `db:"Master_last_event_time" json:"masterLastEventTime"`
-	SlaveLastEventTime       sql.NullString `db:"Slave_last_event_time" json:"slaveLastEventTime"`
-	MasterSlaveTimeDiff      sql.NullInt64  `db:"Master_Slave_time_diff" json:"masterSlaveTimeDiff"`
-	LastIOErrno              sql.NullString `db:"Last_IO_Errno" json:"lastIoErrno"`
-	LastIOError              sql.NullString `db:"Last_IO_Error" json:"lastIoError"`
-	LastSQLErrno             sql.NullString `db:"Last_SQL_Errno" json:"lastSqlErrno"`
-	LastSQLError             sql.NullString `db:"Last_SQL_Error" json:"lastSqlError"`
-	MasterServerID           uint64         `db:"Master_Server_Id" json:"masterServerId"`
-	UsingGtid                sql.NullString `db:"Using_Gtid" json:"usingGtid"`
-	GtidIOPos                sql.NullString `db:"Gtid_IO_Pos" json:"gtidIoPos"`
-	GtidSlavePos             sql.NullString `db:"Gtid_Slave_Pos" json:"gtidSlavePos"`
-	SlaveHeartbeatPeriod     float64        `db:"Slave_Heartbeat_Period" json:"slaveHeartbeatPeriod"`
-	ExecutedGtidSet          sql.NullString `db:"Executed_Gtid_Set" json:"executedGtidSet"`
-	RetrievedGtidSet         sql.NullString `db:"Retrieved_Gtid_Set" json:"retrievedGtidSet"`
-	SlaveSQLRunningState     sql.NullString `db:"Slave_SQL_Running_State" json:"slaveSQLRunningState"`
-	PGExternalID             sql.NullString `db:"external_id" json:"postgresExternalId"`
-	DoDomainIds              sql.NullString `db:"Replicate_Do_Domain_Ids" json:"replicateDoDomainIds"`
-	IgnoreDomainIds          sql.NullString `db:"Replicate_Ignore_Domain_Ids" json:"replicateIgnoreDomainIds"`
-	IgnoreServerIds          sql.NullString `db:"Replicate_Ignore_Server_Ids" json:"replicateIgnoreServerIds"`
-	ReplicateDoDB            sql.NullString `db:"Replicate_Do_DB" json:"replicateDoDb"`
-	ReplicateIgnoreDB        sql.NullString `db:"Replicate_Ignore_DB" json:"replicateIgnoreDb"`
-	ReplicateDoTable         sql.NullString `db:"Replicate_Do_Table" json:"replicateDoTable"`
-	ReplicateIgnoreTable     sql.NullString `db:"Replicate_Ignore_Table" json:"replicateIgnoreTable"`
-	ReplicateWildDoTable     sql.NullString `db:"Replicate_Do_Wild_Table" json:"replicateWildDoTable"`
-	ReplicateWildIgnoreTable sql.NullString `db:"Replicate_Wild_Ignore_Table" json:"replicateWildIgnoreTable"`
-	SQLDelay                 sql.NullInt64  `db:"SQL_Delay" json:"sqlDelay"`
-	SQLRemainingDelay        sql.NullInt64  `db:"SQL_Remaining_Delay" json:"sqlRemainingDelay"`
-	AutoPosition             int            `db:"Auto_Position" json:"autoPosition"`
-	MasterRetryCount         sql.NullInt64  `db:"Master_Retry_Count" json:"masterRetryCount"`
-}
-
-func (s *SlaveStatus) ImportFromReplicaStatus(rs *ReplicaStatus) {
-	s.ConnectionName = sql.NullString{String: rs.ChannelName.String, Valid: rs.ChannelName.Valid}
-	s.ChannelName = sql.NullString{String: rs.ChannelName.String, Valid: rs.ChannelName.Valid}
-	s.MasterHost = rs.SourceHost
-	s.MasterUser = rs.SourceUser
-	s.MasterPort = rs.SourcePort
-	s.MasterLogFile = rs.SourceLogFile
-	s.ReadMasterLogPos = rs.ReadSourceLogPos
-	s.RelayMasterLogFile = rs.RelaySourceLogFile
-	s.SlaveIORunning = rs.ReplicaIORunning
-	s.SlaveSQLRunning = rs.ReplicaSQLRunning
-	s.ExecMasterLogPos = rs.ExecSourceLogPos
-	s.SecondsBehindMaster = sql.NullInt64{Int64: rs.SecondsBehindSource.Int64, Valid: rs.SecondsBehindSource.Valid}
-	s.LastIOErrno = rs.LastIOErrno
-	s.LastIOError = rs.LastIOError
-	s.LastSQLErrno = rs.LastSQLErrno
-	s.LastSQLError = rs.LastSQLError
-	s.MasterServerID = rs.SourceServerID
-	s.UsingGtid = rs.ExecutedGtidSet
-	s.GtidIOPos = rs.RetrievedGtidSet
-	s.GtidSlavePos = rs.RetrievedGtidSet
-	s.ExecutedGtidSet = rs.ExecutedGtidSet
-	s.RetrievedGtidSet = rs.RetrievedGtidSet
-	s.SlaveSQLRunningState = rs.ReplicaSQLRunningState
-	s.PGExternalID = rs.SourceUUID
-	s.DoDomainIds = rs.ReplicateDoDB
-	s.IgnoreDomainIds = rs.ReplicateIgnoreDB
-	s.IgnoreServerIds = rs.ReplicateIgnoreServerIds
-	s.ReplicateDoDB = rs.ReplicateDoDB
-	s.ReplicateIgnoreDB = rs.ReplicateIgnoreDB
-	s.ReplicateDoTable = rs.ReplicateDoTable
-	s.ReplicateIgnoreTable = rs.ReplicateIgnoreTable
-	s.ReplicateWildDoTable = rs.ReplicateWildDoTable
-	s.ReplicateWildIgnoreTable = rs.ReplicateWildIgnoreTable
-	s.SQLDelay = rs.SQLDelay
-	s.SQLRemainingDelay = rs.SQLRemainingDelay
-	s.AutoPosition = rs.AutoPosition
-	s.MasterRetryCount = rs.SourceRetryCount
-}
-
-type ReplicaStatus struct {
-	ReplicaIOState                 sql.NullString `db:"Replica_IO_State" json:"replicaIoState"`
-	SourceHost                     sql.NullString `db:"Source_Host" json:"sourceHost"`
-	SourceUser                     sql.NullString `db:"Source_User" json:"sourceUser"`
-	SourcePort                     sql.NullString `db:"Source_Port" json:"sourcePort"`
-	ConnectRetry                   sql.NullInt64  `db:"Connect_Retry" json:"connectRetry"`
-	SourceLogFile                  sql.NullString `db:"Source_Log_File" json:"sourceLogFile"`
-	ReadSourceLogPos               sql.NullString `db:"Read_Source_Log_Pos" json:"readSourceLogPos"`
-	RelayLogFile                   sql.NullString `db:"Relay_Log_File" json:"relayLogFile"`
-	RelayLogPos                    sql.NullString `db:"Relay_Log_Pos" json:"relayLogPos"`
-	RelaySourceLogFile             sql.NullString `db:"Relay_Source_Log_File" json:"relaySourceLogFile"`
-	ReplicaIORunning               sql.NullString `db:"Replica_IO_Running" json:"replicaIoRunning"`
-	ReplicaSQLRunning              sql.NullString `db:"Replica_SQL_Running" json:"replicaSqlRunning"`
-	ReplicateDoDB                  sql.NullString `db:"Replicate_Do_DB" json:"replicateDoDb"`
-	ReplicateIgnoreDB              sql.NullString `db:"Replicate_Ignore_DB" json:"replicateIgnoreDb"`
-	ReplicateDoTable               sql.NullString `db:"Replicate_Do_Table" json:"replicateDoTable"`
-	ReplicateIgnoreTable           sql.NullString `db:"Replicate_Ignore_Table" json:"replicateIgnoreTable"`
-	ReplicateWildDoTable           sql.NullString `db:"Replicate_Wild_Do_Table" json:"replicateWildDoTable"`
-	ReplicateWildIgnoreTable       sql.NullString `db:"Replicate_Wild_Ignore_Table" json:"replicateWildIgnoreTable"`
-	LastErrno                      sql.NullString `db:"Last_Errno" json:"lastErrno"`
-	LastError                      sql.NullString `db:"Last_Error" json:"lastError"`
-	SkipCounter                    sql.NullInt64  `db:"Skip_Counter" json:"skipCounter"`
-	ExecSourceLogPos               sql.NullString `db:"Exec_Source_Log_Pos" json:"execSourceLogPos"`
-	RelayLogSpace                  sql.NullInt64  `db:"Relay_Log_Space" json:"relayLogSpace"`
-	UntilCondition                 sql.NullString `db:"Until_Condition" json:"untilCondition"`
-	UntilLogFile                   sql.NullString `db:"Until_Log_File" json:"untilLogFile"`
-	UntilLogPos                    sql.NullString `db:"Until_Log_Pos" json:"untilLogPos"`
-	SourceSSLAllowed               sql.NullString `db:"Source_SSL_Allowed" json:"sourceSslAllowed"`
-	SourceSSLCaFile                sql.NullString `db:"Source_SSL_CA_File" json:"sourceSslCaFile"`
-	SourceSSLCaPath                sql.NullString `db:"Source_SSL_CA_Path" json:"sourceSslCaPath"`
-	SourceSSLCert                  sql.NullString `db:"Source_SSL_Cert" json:"sourceSslCert"`
-	SourceSSLCipher                sql.NullString `db:"Source_SSL_Cipher" json:"sourceSslCipher"`
-	SourceSSLKey                   sql.NullString `db:"Source_SSL_Key" json:"sourceSslKey"`
-	SecondsBehindSource            sql.NullInt64  `db:"Seconds_Behind_Source" json:"secondsBehindSource"`
-	SourceSSLVerifyServerCert      sql.NullString `db:"Source_SSL_Verify_Server_Cert" json:"sourceSslVerifyServerCert"`
-	LastIOErrno                    sql.NullString `db:"Last_IO_Errno" json:"lastIoErrno"`
-	LastIOError                    sql.NullString `db:"Last_IO_Error" json:"lastIoError"`
-	LastSQLErrno                   sql.NullString `db:"Last_SQL_Errno" json:"lastSqlErrno"`
-	LastSQLError                   sql.NullString `db:"Last_SQL_Error" json:"lastSqlError"`
-	ReplicateIgnoreServerIds       sql.NullString `db:"Replicate_Ignore_Server_Ids" json:"replicateIgnoreServerIds"`
-	SourceServerID                 uint64         `db:"Source_Server_Id" json:"sourceServerId"`
-	SourceUUID                     sql.NullString `db:"Source_UUID" json:"sourceUuid"`
-	SourceInfoFile                 sql.NullString `db:"Source_Info_File" json:"sourceInfoFile"`
-	SQLDelay                       sql.NullInt64  `db:"SQL_Delay" json:"sqlDelay"`
-	SQLRemainingDelay              sql.NullInt64  `db:"SQL_Remaining_Delay" json:"sqlRemainingDelay"`
-	ReplicaSQLRunningState         sql.NullString `db:"Replica_SQL_Running_State" json:"replicaSqlRunningState"`
-	SourceRetryCount               sql.NullInt64  `db:"Source_Retry_Count" json:"sourceRetryCount"`
-	SourceBind                     sql.NullString `db:"Source_Bind" json:"sourceBind"`
-	LastIOErrorTimestamp           sql.NullString `db:"Last_IO_Error_Timestamp" json:"lastIoErrorTimestamp"`
-	LastSQLExceptionErrorTimestamp sql.NullString `db:"Last_SQL_Error_Timestamp" json:"lastSqlErrorTimestamp"`
-	SourceSSLCrl                   sql.NullString `db:"Source_SSL_Crl" json:"sourceSslCrl"`
-	SourceSSLCrlpath               sql.NullString `db:"Source_SSL_Crlpath" json:"sourceSslCrlpath"`
-	RetrievedGtidSet               sql.NullString `db:"Retrieved_Gtid_Set" json:"retrievedGtidSet"`
-	ExecutedGtidSet                sql.NullString `db:"Executed_Gtid_Set" json:"executedGtidSet"`
-	AutoPosition                   int            `db:"Auto_Position" json:"autoPosition"`
-	ReplicateRewriteDB             sql.NullString `db:"Replicate_Rewrite_DB" json:"replicateRewriteDb"`
-	ChannelName                    sql.NullString `db:"Channel_Name" json:"channelName"`
-	SourceTLSVersion               sql.NullString `db:"Source_TLS_Version" json:"sourceTlsVersion"`
-	SourcePublicKeyPath            sql.NullString `db:"Source_public_key_path" json:"sourcePublicKeyPath"`
-	GetSourcePublicKey             sql.NullString `db:"Get_Source_public_key" json:"getSourcePublicKey"`
-	NetworkNamespace               sql.NullString `db:"Network_Namespace" json:"networkNamespace"`
-}
-
-type Privileges struct {
-	Select_priv      string `json:"selectPriv"`
-	Process_priv     string `json:"processPriv"`
-	Super_priv       string `json:"superPriv"`
-	Repl_slave_priv  string `json:"replSlavePriv"`
-	Repl_client_priv string `json:"replClientPriv"`
-	Reload_priv      string `json:"reloadPriv"`
-}
-
-type SpiderTableNoSync struct {
-	Tbl_src      string
-	Tbl_src_link string
-	Tbl_dest     string
-	Srv_dsync    string
-	Srv_sync     string
-}
-
-type BinlogEvents struct {
-	Log_name    string `db:"Log_name" json:"logName"`
-	Pos         uint   `db:"Pos" json:"pos"`
-	Event_type  string `db:"Event_type" json:"eventType"`
-	Server_id   uint   `db:"Server_id" json:"serverId"`
-	End_log_pos uint   `db:"End_log_pos" json:"endLogPos"`
-	Info        string `db:"Info" json:"info"`
-}
-
-type MySQLServer struct {
-	Server_name string `db:"Server_name" json:"serverName"`
-	Host        string `db:"Host" json:"host"`
-	Db          string `db:"Db" json:"db"`
-	Username    string `db:"Username" json:"username"`
-	Password    string `db:"Password" json:"password"`
-	Port        uint   `db:"Port" json:"port"`
-	Socket      string `db:"Socket" json:"socket"`
-	Wrapper     string `db:"Wrapper" json:"wrapper"`
-	Owner       string `db:"Owner" json:"owner"`
-}
-
-type Variable struct {
-	Variable_name string `json:"variableName"`
-	Value         string `json:"value"`
-}
-
-type Binarylogs struct {
-	Log_name  string `json:"logName"`
-	File_size uint   `json:"fileSize"`
-	Encrypted string `json:"encrypted"` //mysql 8.0
-}
-
-type Explain struct {
-	Id            uint           `db:"id" json:"id"`
-	Select_type   sql.NullString `db:"select_type" json:"selectType"`
-	Table         sql.NullString `db:"table" json:"table"`
-	Type          sql.NullString `db:"type" json:"type"`
-	Possible_keys sql.NullString `db:"possible_keys" json:"possibleKeys"`
-	Key           sql.NullString `db:"key" json:"key"`
-	Key_len       sql.NullString `db:"key_len" json:"keyLen"`
-	Ref           sql.NullString `db:"ref" json:"ref"`
-	Rows          sql.NullString `db:"rows" json:"rows"`
-	Extra         sql.NullString `db:"Extra" json:"extra"`
-}
-
-type VariableSorter []Variable
-
-func (a VariableSorter) Len() int           { return len(a) }
-func (a VariableSorter) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
-func (a VariableSorter) Less(i, j int) bool { return a[i].Variable_name < a[j].Variable_name }
-
-func MySQLConnect(user string, password string, address string, parameters ...string) (*sqlx.DB, error) {
-	dsn := user + ":" + password + "@" + address + "/"
-	if len(parameters) > 0 {
-		dsn += "?" + parameters[0]
-	}
-	db, err := sqlx.Connect("mysql", dsn)
-	return db, err
-}
-
-// SQLiteConnect returns a SQLite connection
-func SQLiteConnect(path string) (*sqlx.DB, error) {
-	db, err := sqlx.Connect("sqlite3", path+"/arbitrator.db")
-	return db, err
-}
-
-func GetQueryDigest(q string) string {
-	f := query.Fingerprint(q)
-	return f
-}
-
-func GetAddress(host string, port string, socket string) string {
-	var address string
-	if host != "" {
-		address = "tcp(" + host + ":" + port + ")"
-	} else {
-		address = "unix(" + socket + ")"
-	}
-	return address
-}
-
-func GetQueryExplain(db *sqlx.DB, version *version.Version, schema string, query string) ([]Explain, string, error) {
-	pl := []Explain{}
-	var err error
-	if schema != "" {
-		_, err = db.Exec("USE " + schema)
-	}
-	stmt := "Explain " + query
-	err = db.Select(&pl, stmt)
-	if err != nil {
-		return nil, stmt, fmt.Errorf("ERROR: Could not get Explain: %s", err)
-	}
-	return pl, stmt, nil
-}
-
-func GetMetaDataLock(db *sqlx.DB, version *version.Version) ([]MetaDataLock, string, error) {
-	/*	select pid from pg_locks l
-		join pg_class t on l.relation = t.oid
-		and t.relkind = 'r'  */
-	pl := []MetaDataLock{}
-	var err error
-	query := "SELECT * FROM information_schema.metadata_lock_info"
-	if version.IsMariaDB() {
-		//MariaDB
-		err = db.Select(&pl, query)
-	}
-	if err != nil {
-		return nil, query, fmt.Errorf("ERROR: Could not get MetaDataLock: %s", err)
-	}
-	return pl, query, nil
-}
-
-func GetQueryResponseTime(db *sqlx.DB, version *version.Version) ([]ResponseTime, string, error) {
-	pl := []ResponseTime{}
-	var err error
-	stmt := "SELECT * FROM INFORMATION_SCHEMA.QUERY_RESPONSE_TIME"
-	if version.IsMySQL() || version.IsPostgreSQL() {
-		return nil, stmt, fmt.Errorf("ERROR: QUERY_RESPONSE_TIME not available on MySQL or PostgeSQL: %s", err)
-	}
-	err = db.Select(&pl, stmt)
-	if err != nil {
-		return nil, stmt, fmt.Errorf("ERROR: Could not get query response time: %s", err)
-	}
-	return pl, stmt, nil
-}
-
-func GetBinaryLogs(db *sqlx.DB, version *version.Version, metamap *BinaryLogMetaMap) (int, string, []string, string, error) {
-	counter := 0
-	oldest := ""
-	trimmed := make([]string, 0)
-	query := "SHOW BINARY LOGS"
-	if version.IsPostgreSQL() {
-		return counter, oldest, trimmed, query, fmt.Errorf("ERROR: QUERY_RESPONSE_TIME not available on PostgeSQL")
-	}
-	rows, err := db.Queryx(query)
-
-	if err != nil {
-		return counter, oldest, trimmed, query, errors.New("Could not get binary logs: " + err.Error())
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v Binarylogs
-		if version.IsMySQLOrPercona() && version.Major >= 8 {
-			err = rows.Scan(&v.Log_name, &v.File_size, &v.Encrypted)
-		} else {
-			err = rows.Scan(&v.Log_name, &v.File_size)
-		}
-		if err != nil {
-			return counter, oldest, trimmed, query, errors.New("Could not get binary logs: " + err.Error())
-		}
-		if oldest == "" {
-			oldest = v.Log_name
-		}
-		if meta, exists := metamap.LoadOrStore(v.Log_name, &BinaryLogMetadata{Filename: v.Log_name, Size: v.File_size}); exists {
-			if meta.Size != v.File_size {
-				meta.Size = v.File_size
-			}
-		}
-		counter++
-	}
-
-	trimmed = metamap.ClearObsoleteMetadata(oldest)
-
-	return counter, oldest, trimmed, query, nil
-}
-
-func AnalyzeQuery(db *sqlx.DB, version *version.Version, schema string, query string) (string, string, error) {
-	var res string
-	if schema != "" {
-		db.Exec("USE " + schema)
-	}
-	stmt := "ANALYZE  FORMAT=JSON " + query
-	rows, err := db.Query(stmt)
-	if err != nil {
-		return "", stmt, err
-	}
-	defer rows.Close()
-
-	if rows.Next() {
-		if err := rows.Scan(&res); err != nil {
-			return res, stmt, err
-		}
-	}
-	return res, stmt, err
-}
-
-func GetProcesslistTable(db *sqlx.DB, version *version.Version, inactive_querying bool, order_by_trx_time bool, full_process_is bool, limit string, user string) ([]Processlist, string, error) {
-	pl := []Processlist{}
-	var err error
-	stmt := ""
-	filter_order_limit := ""
-	filter_user := "1=1"
-
-	if user != "" {
-		filter_user = "  User='" + user + "' "
-	}
-	if inactive_querying || order_by_trx_time {
-		filter_order_limit = " WHERE " + filter_user
-		if order_by_trx_time {
-			filter_order_limit += " AND  (a.command='query' or b.trx_isolation_level IS NOT NULL)"
-		}
-	} else {
-		filter_order_limit = " WHERE " + filter_user + " AND  a.command='query' "
-	}
-	if order_by_trx_time {
-		if version.IsMariaDB() || (version.IsPercona() && version.GreaterEqual("8.0")) {
-			filter_order_limit += " ORDER BY IF(a.Command='Query',a.TIME_MS, b.trx_started) DESC LIMIT " + limit
-		} else {
-			filter_order_limit += " ORDER BY IF(a.Command='Query',a.TIME, b.trx_started) DESC LIMIT " + limit
-		}
-	} else {
-		if version.IsMariaDB() || (version.IsPercona() && version.GreaterEqual("8.0")) {
-			filter_order_limit += " ORDER BY a.TIME_MS DESC LIMIT " + limit
-		} else {
-			filter_order_limit += " ORDER BY a.TIME DESC LIMIT " + limit
-		}
-	}
-	if version.IsMariaDB() {
-		//MariaDB
-		stmt = "SELECT " +
-			"a.Id, " +
-			"a.User, " +
-			"a.Host, " +
-			"a.`Db` AS `db`, " +
-			"IF(a.Command='Sleep' AND b.trx_isolation_level IS NOT NULL, 'Trx sleep', a.Command) as Command, " +
-			"a.Time_ms/1000 as Time, " +
-			"a.State, " +
-			"SUBSTRING(COALESCE(a.INFO_BINARY,''),1,1000) as Info, " +
-			"CASE WHEN a.Max_Stage < 2 THEN Progress ELSE (a.Stage-1)/a.Max_Stage*100+a.Progress/a.Max_Stage END AS Progress, " +
-			"GREATEST(COALESCE(TIMESTAMPDIFF(SECOND,b.trx_started, now()),0),0) as trx_time, " +
-			"COALESCE(b.trx_isolation_level,'') as trx_isolation_level, " +
-			"COALESCE(b.trx_tables_in_use,0) as trx_tables_in_use, " +
-			"COALESCE(b.trx_tables_locked,0) as trx_tables_locked, " +
-			"COALESCE(b.trx_lock_structs,0) as trx_lock_structs, " +
-			"COALESCE(b.trx_lock_memory_bytes,0) as trx_lock_memory_bytes, " +
-			"COALESCE(b.trx_rows_locked,0) as trx_rows_locked, " +
-			"COALESCE(b.trx_rows_modified,0) as trx_rows_modified, " +
-			"COALESCE(b.trx_is_read_only,0) as trx_is_read_only " +
-			"FROM INFORMATION_SCHEMA.PROCESSLIST a " +
-			"LEFT JOIN INFORMATION_SCHEMA.INNODB_TRX b ON b.trx_mysql_thread_id=a.id " +
-			filter_order_limit
-		if !full_process_is {
-			stmt = "SHOW FULL PROCESSLIST"
-		}
-	} else if version.IsMySQLOrPercona() {
-		//MySQL
-		stmt = "SELECT " +
-			"a.Id, " +
-			"a.User, " +
-			"a.Host, " +
-			"a.`Db` AS `db`, " +
-			"IF(a.Command='Sleep' AND b.trx_isolation_level IS NOT NULL, 'Trx sleep', a.Command) as Command, " +
-			"a.Time as Time, " +
-			"a.State, " +
-			"SUBSTRING(COALESCE(a.INFO,''),1,1000) as Info, " +
-			"0 as Progress, " +
-			"GREATEST(COALESCE(TIMESTAMPDIFF(SECOND,b.trx_started, now()),0),0) as trx_time, " +
-			"COALESCE(b.trx_isolation_level,'') as trx_isolation_level, " +
-			"COALESCE(b.trx_tables_in_use,0) as trx_tables_in_use, " +
-			"COALESCE(b.trx_tables_locked,0) as trx_tables_locked, " +
-			"COALESCE(b.trx_lock_structs,0) as trx_lock_structs, " +
-			"COALESCE(b.trx_lock_memory_bytes,0) as trx_lock_memory_bytes, " +
-			"COALESCE(b.trx_rows_locked,0) as trx_rows_locked, " +
-			"COALESCE(b.trx_rows_modified,0) as trx_rows_modified, " +
-			"COALESCE(b.trx_is_read_only,0) as trx_is_read_only " +
-			"FROM INFORMATION_SCHEMA.PROCESSLIST a " +
-			"LEFT JOIN INFORMATION_SCHEMA.INNODB_TRX b ON b.trx_mysql_thread_id=a.id " +
-			filter_order_limit
-		if version.IsPercona() && version.GreaterEqual("8.0") {
-			stmt = "SELECT " +
-				"a.Id, " +
-				"a.User, " +
-				"a.Host, " +
-				"a.`Db` AS `db`, " +
-				"IF(a.Command='Sleep' AND b.trx_isolation_level IS NOT NULL, 'Trx sleep', a.Command) as Command, " +
-				"a.Time_ms as Time, " +
-			"a.State, " +
-			"SUBSTRING(COALESCE(a.INFO,''),1,1000) as Info, " +
-			"0 as Progress, " +
-			"GREATEST(COALESCE(TIMESTAMPDIFF(SECOND,b.trx_started, now()),0),0) as trx_time, " +
-			"COALESCE(b.trx_isolation_level,'') as trx_isolation_level, " +
-				"COALESCE(b.trx_tables_in_use,0) as trx_tables_in_use, " +
-				"COALESCE(b.trx_tables_locked,0) as trx_tables_locked, " +
-				"COALESCE(b.trx_lock_structs,0) as trx_lock_structs, " +
-				"COALESCE(b.trx_lock_memory_bytes,0) as trx_lock_memory_bytes, " +
-				"COALESCE(b.trx_rows_locked,0) as trx_rows_locked, " +
-				"COALESCE(b.trx_rows_modified,0) as trx_rows_modified, " +
-				"COALESCE(b.trx_is_read_only,0) as trx_is_read_only " +
-				"FROM INFORMATION_SCHEMA.PROCESSLIST a " +
-				"LEFT JOIN INFORMATION_SCHEMA.INNODB_TRX b ON b.trx_mysql_thread_id=a.id " +
-				filter_order_limit
-		}
-		if !full_process_is {
-			stmt = "SHOW FULL PROCESSLIST"
-		}
-	} else if version.IsPostgreSQL() {
-		// WHERE state <> 'idle' 		AND pid<>pg_backend_pid()
-		stmt = `SELECT pid as "Id", coalesce(usename,'') as "User",coalesce(client_hostname || client_port,'') as "Host" , coalesce(datname,'') as db , coalesce(query,'') as "Command", extract(epoch from NOW()) - extract(epoch from query_start) as "Time",  coalesce(state,'') as "State",COALESCE(application_name,'')  as "Info" ,0 as "Progress"  FROM pg_stat_activity`
-	}
-	err = db.Select(&pl, stmt)
-	if err != nil {
-		return nil, stmt, fmt.Errorf("ERROR: Could not get processlist: %s", err)
-	}
-	return pl, stmt, nil
-}
-
-/*func GetProcesslistTableFromUser(db *sqlx.DB, version *version.Version, user string) ([]Processlist, string, error) {
-	pl := []Processlist{}
-	var err error
-	stmt := ""
-	if version.IsMariaDB() {
-		//MariaDB
-		stmt = "SELECT Id, User, Host, `Db` AS `db`, Command, Time_ms as Time, State, SUBSTRING(COALESCE(INFO_BINARY,''),1,1000) as Info, CASE WHEN Max_Stage < 2 THEN Progress ELSE (Stage-1)/Max_Stage*100+Progress/Max_Stage END AS Progress FROM INFORMATION_SCHEMA.PROCESSLIST WHERE User='+ user +'"
-	} else if version.IsMySQLOrPercona() {
-		//MySQL
-		stmt = "SELECT Id, User, Host, `Db` AS `db`, Command, Time as Time, State, SUBSTRING(COALESCE(INFO,''),1,1000) as Info ,0 as Progress FROM INFORMATION_SCHEMA.PROCESSLIST WHERE  User='+ user +'"
-		if version.GreaterEqual("8.0") {
-			stmt = "SELECT Id, User, Host, `Db` AS `db`, Command, Time_ms as Time, State, SUBSTRING(COALESCE(INFO,''),1,1000) as Info ,0 as Progress FROM INFORMATION_SCHEMA.PROCESSLIST WHERE   User='+ user +'"
-		}
-	} else if version.IsPostgreSQL() {
-		// WHERE state <> 'idle' 		AND pid<>pg_backend_pid()
-		stmt = `SELECT pid as "Id", coalesce(usename,'') as "User",coalesce(client_hostname || client_port,'') as "Host" , coalesce(datname,'') as db , coalesce(query,'') as "Command", extract(epoch from NOW()) - extract(epoch from query_start) as "Time",  coalesce(state,'') as "State",COALESCE(application_name,'')  as "Info" ,0 as "Progress"  FROM pg_stat_activity`
-	}
-	err = db.Select(&pl, stmt)
-	if err != nil {
-		return nil, stmt, fmt.Errorf("ERROR: Could not get processlist: %s", err)
-	}
-	return pl, stmt, nil
-}
-
-func GetProcesslist(db *sqlx.DB, version *version.Version) ([]Processlist, string, error) {
-	pl := []Processlist{}
-	var err error
-	query := ""
-	if version.IsMariaDB() {
-		//MariaDB
-		query = "SHOW FULL PROCESSLIST"
-	} else if version.IsPostgreSQL() {
-		// WHERE state <> 'idle' 		AND pid<>pg_backend_pid()
-		query = `SELECT pid as "Id", coalesce(usename,'') as "User",coalesce(client_hostname || client_port,'') as "Host" , coalesce(datname,'') as db ,COALESCE(application_name,'')  as "Command", extract(epoch from NOW()) - extract(epoch from query_start) as "Time",  coalesce(state,'') as "State", coalesce(query,'')  as "Info" ,0 as "Progress"  FROM pg_stat_activity`
-	} else {
-		//MySQL
-		query = "SHOW FULL PROCESSLIST"
-	}
-	err = db.Select(&pl, query)
-	if err != nil {
-		return nil, query, fmt.Errorf("ERROR: Could not get processlist: %s", err)
-	}
-	return pl, query, nil
-}*/
-
-func GetServers(db *sqlx.DB) ([]MySQLServer, string, error) {
-	db.MapperFunc(strings.Title)
-	var err error
-	ss := []MySQLServer{}
-	query := "SELECT * FROM mysql.servers"
-	err = db.Select(&ss, query)
-	return ss, query, err
-}
-
-func GetLastPseudoGTID(db *sqlx.DB) (string, string, error) {
-	var value string
-	value = ""
-	query := "select * from replication_manager_schema.pseudo_gtid_v"
-	err := db.QueryRowx(query).Scan(&value)
-	return value, query, err
-}
-
-func GetBinlogEventPseudoGTID(db *sqlx.DB, uuid string, lastfile string) (string, string, string, error) {
-
-	lastpos := "4"
-	exitloop := true
-	logs := ""
-	for exitloop {
-		events := []BinlogEvents{}
-		sql := "show binlog events IN '" + lastfile + "'  from " + lastpos + " LIMIT 60"
-		logs += sql + "\n"
-		err := db.Select(&events, sql)
-		if err != nil {
-			return "", "", logs, err
-		}
-
-		for _, row := range events {
-			pos := strconv.FormatUint(uint64(row.Pos), 10)
-			endpos := strconv.FormatUint(uint64(row.End_log_pos), 10)
-			if strings.Contains(row.Info, uuid) {
-				return row.Log_name, pos, logs, err
-			}
-			lastpos = endpos
-		}
-		if len(events) == 0 {
-			binlogindex, _ := strconv.Atoi(strings.Split(lastfile, ".")[1])
-			binlogindex = binlogindex - 1
-			lastfile = strings.Split(lastfile, ".")[0] + "." + fmt.Sprintf("%06d", binlogindex)
-			lastpos = "4"
-		}
-	}
-	return "", "", logs, errors.New("Not found Psudo GTID")
-}
-
-func GetBinlogPosAfterSkipNumberOfEvents(db *sqlx.DB, file string, pos string, skip int) (string, string, string, error) {
-
-	events := []BinlogEvents{}
-	sql := "show binlog events IN '" + file + "'  from " + pos + " LIMIT " + strconv.Itoa(skip)
-
-	err := db.Select(&events, sql)
-	if err != nil {
-		return "", "", sql, err
-	}
-	if len(events) == 0 {
-		return "", "", sql, err
-	}
-	return events[(len(events) - 1)].Log_name, strconv.FormatUint(uint64(events[(len(events)-1)].Pos), 10), sql, err
-}
-
-func GetNumberOfEventsAfterPos(db *sqlx.DB, lastfile string, lastpos string) (int, string, error) {
-
-	exitloop := true
-	logs := ""
-	ct := 0
-	for exitloop {
-		events := []BinlogEvents{}
-		sql := "show binlog events IN '" + lastfile + "'  from " + lastpos + " LIMIT 1"
-		logs += sql + "\n"
-		err := db.Select(&events, sql)
-		if err != nil {
-			return 0, logs, err
-		}
-
-		for _, row := range events {
-			lastfile = strconv.FormatUint(uint64(row.End_log_pos), 10)
-		}
-		if len(events) == 0 {
-			return ct, logs, nil
-		}
-		ct = ct + 1
-	}
-	return 0, logs, errors.New("Not found Psudo GTID")
-}
-
-func GetMaxscaleVersion(db *sqlx.DB) (string, error) {
-	var value string
-	value = ""
-	err := db.QueryRowx("Select @@maxscale_version").Scan(&value)
-	return value, err
-}
-
-type ChangeMasterOpt struct {
-	Host      string
-	Port      string
-	User      string
-	Password  string
-	Retry     string
-	Heartbeat string
-	SSL       bool
-	Logfile   string
-	Logpos    string
-	Mode      string
-
-	RetryCount string // Start from MariaDB 12 and MySQL 8.4
-
-	Channel         string
-	PostgressDB     string
-	IsDelayed       bool
-	Delay           string
-	DoDomainIds     string
-	IgnoreDomainIds string
-	IgnoreServerIds string
-	//	SSLCa     string
-	//	SSLCert   string
-	//	SSLKey    string
-}
-
-func ChangeReplicationPassword(db *sqlx.DB, opt ChangeMasterOpt, myver *version.Version) (string, error) {
-	_, err := StopSlave(db, opt.Channel, myver)
-	if err != nil {
-		return "Stop slave error", err
-	}
-	masterOrSource := "MASTER"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 23)) {
-		masterOrSource = "SOURCE"
-	}
-	cm := ""
-	if myver.IsMariaDB() && opt.Channel != "" {
-		cm += "CHANGE " + masterOrSource + " '" + opt.Channel + "' TO "
-	} else {
-		cm += "CHANGE  " + masterOrSource + " TO "
-	}
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 23)) {
-		cm = "CHANGE REPLICATION SOURCE TO "
-	}
-
-	if opt.Mode == "GROUP_REPL" {
-		cm += masterOrSource + "_user='" + opt.User + "', " + masterOrSource + "_password='" + opt.Password + "'"
-	} else {
-		cm += " " + masterOrSource + "_user='" + opt.User + "', " + masterOrSource + "_password='" + opt.Password + "'"
-	}
-
-	if myver.IsMySQLOrPercona() && opt.Channel != "" {
-		cm += " FOR CHANNEL '" + opt.Channel + "'"
-	}
-	_, err = db.Exec(cm)
-	cm = strings.Replace(cm, opt.Password, "XXX", -1)
-	if err != nil {
-		return cm, fmt.Errorf("Change "+masterOrSource+" statement %s failed, reason: %s", cm, err)
-	}
-	_, err = StartSlave(db, opt.Channel, myver)
-	if err != nil {
-		return "Start slave error", err
-	}
-	return cm, nil
-}
-
-func ChangeMaster(db *sqlx.DB, opt ChangeMasterOpt, myver *version.Version) (string, error) {
-	//CREATE PUBLICATION alltables FOR ALL TABLES;
-	/*
-		Group replication we will check opt.Mode=GROUP_REPL
-		The master_host is not used
-		mysql> CHANGE MASTER TO MASTER_USER='rpl_user', MASTER_PASSWORD='password' \\
-				      FOR CHANNEL 'group_replication_recovery';
-
-		Or from MySQL 8.0.23:
-		mysql> CHANGE REPLICATION SOURCE TO SOURCE_USER='rpl_user', SOURCE_PASSWORD='password' \\
-				      FOR CHANNEL 'group_replication_recovery';
-	*/
-	masterOrSource := "MASTER"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 23)) {
-		masterOrSource = "SOURCE"
-	}
-	cm := ""
-	if myver.IsPostgreSQL() {
-		if opt.Channel == "" {
-			opt.Channel = "alltables"
-		}
-		cm += "CREATE SUBSCRIPTION " + opt.Channel + " CONNECTION 'dbname=" + opt.PostgressDB + " host=" + misc.Unbracket(opt.Host) + " user=" + opt.User + " port=" + opt.Port + " password=" + opt.Password + " ' PUBLICATION  " + opt.Channel + " WITH (enabled=false, copy_data=false, create_slot=true)"
-	} else {
-		if myver.IsMariaDB() && opt.Channel != "" {
-			cm += "CHANGE " + masterOrSource + " '" + opt.Channel + "' TO "
-		} else {
-			cm += "CHANGE  " + masterOrSource + " TO "
-		}
-		if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 23)) {
-			cm = "CHANGE REPLICATION SOURCE TO "
-		}
-
-		if opt.Mode == "GROUP_REPL" {
-			cm += masterOrSource + "_user='" + opt.User + "', " + masterOrSource + "_password='" + opt.Password + "'"
-		} else {
-			cm += " " + masterOrSource + "_host='" + misc.Unbracket(opt.Host) + "', " + masterOrSource + "_port=" + opt.Port + ", " + masterOrSource + "_user='" + opt.User + "', " + masterOrSource + "_password='" + opt.Password + "', " + masterOrSource + "_connect_retry=" + opt.Retry + ", " + masterOrSource + "_heartbeat_period=" + opt.Heartbeat
-		}
-		if opt.IsDelayed {
-			cm += " ," + masterOrSource + "_delay=" + opt.Delay
-		}
-		if myver.IsMariaDB() {
-			if opt.DoDomainIds != "" {
-				cm += " ,DO_DOMAIN_IDS=" + opt.DoDomainIds
-			}
-			if opt.IgnoreDomainIds != "" {
-				cm += " ,IGNORE_DOMAIN_IDS=" + opt.IgnoreDomainIds
-			}
-			if opt.IgnoreDomainIds != "" {
-				cm += " ,IGNORE_DOMAIN_IDS=" + opt.IgnoreDomainIds
-			}
-			if opt.IgnoreServerIds != "" {
-				cm += " ,IGNORE_SERVER_IDS=" + opt.IgnoreServerIds
-			}
-		}
-		switch opt.Mode {
-		case "SLAVE_POS":
-			cm += ", " + masterOrSource + "_USE_GTID=SLAVE_POS"
-		case "CURRENT_POS":
-			if myver.Greater("10.10.0") && myver.IsMariaDB() {
-				cm += ", " + masterOrSource + "_USE_GTID=SLAVE_POS, MASTER_DEMOTE_TO_SLAVE=1"
-			} else {
-				cm += ", " + masterOrSource + "_USE_GTID=CURRENT_POS"
-			}
-		case "MXS":
-			cm += ", " + masterOrSource + "_log_file='" + opt.Logfile + "', " + masterOrSource + "_log_pos=" + opt.Logpos
-		case "POSITIONAL":
-			cm += ", " + masterOrSource + "_log_file='" + opt.Logfile + "', " + masterOrSource + "_log_pos=" + opt.Logpos
-			if myver.IsMariaDB() {
-				cm += ", " + masterOrSource + "_USE_GTID=NO"
-			}
-		case "MASTER_AUTO_POSITION":
-			cm += ", " + masterOrSource + "_AUTO_POSITION=1"
-		}
-		if opt.SSL {
-			cm += ", " + masterOrSource + "_SSL=1"
-			//cm +=, MASTER_SSL_CA='" + opt.SSLCa + "', MASTER_SSL_CERT='" + opt.SSLCert + "', MASTER_SSL_KEY=" + opt.SSLKey + "'"
-		}
-
-		// Retry count supported from MariaDB 12 and MySQL 8.4
-		if myver.IsMariaDBGreater12() || myver.IsMySQLOrPerconaGreater84() {
-			if opt.RetryCount != "" {
-				cm += ", " + masterOrSource + "_RETRY_COUNT=" + opt.RetryCount
-			}
-		}
-		if myver.IsMySQLOrPercona() && opt.Channel != "" {
-			cm += " FOR CHANNEL '" + opt.Channel + "'"
-		}
-	}
-	_, err := db.Exec(cm)
-	cm = strings.Replace(cm, opt.Password, "XXX", -1)
-	if err != nil {
-		return cm, fmt.Errorf("Change "+masterOrSource+" statement %s failed, reason: %s", cm, err)
-	}
-	return cm, nil
-}
-
-func GetCPUUsageFromUserStats(db *sqlx.DB) (string, string, error) {
-	db.MapperFunc(strings.Title)
-	var value string
-	value = ""
-
-	query := "select SUM(BUSY_TIME) FROM INFORMATION_SCHEMA.USER_STATISTICS"
-
-	err := db.QueryRowx(query).Scan(&value)
-	return value, query, err
-}
-
-func MariaDBVersion(server string) int {
-	if server == "" {
-		return 0
-	}
-	re := regexp.MustCompile(`([0-9]+).([0-9]+).([0-9]+)*`)
-	match := re.FindStringSubmatch(server)
-	if len(match[1]) == 0 || len(match[2]) == 0 || len(match[3]) == 0 {
-		return 0
-	}
-	x, _ := strconv.Atoi(match[1])
-	y, _ := strconv.Atoi(match[2])
-	z, _ := strconv.Atoi(match[3])
-	return (x*10000 + y*100 + z)
-	//return ((versionSplit[0]*10000+versionSplit[1])*100 + versionSplit[2])
-}
-
-func GetDBVersion(db *sqlx.DB) (*version.Version, string, error) {
-	stmt := "SELECT version()"
-	var vString string
-	var vComment string
-	err := db.QueryRowx(stmt).Scan(&vString)
-	if err != nil {
-		return &version.Version{}, stmt, err
-	}
-	v, _ := version.NewMySQLVersion(vString, "")
-	if !v.IsPostgreSQL() {
-		stmt = "SELECT @@version_comment"
-		db.QueryRowx(stmt).Scan(&vComment)
-	}
-	nv, _ := version.NewMySQLVersion(vString, vComment)
-	return nv, stmt, nil
-}
-
-// Unused does not look like safe way or documenting it
-func GetHostFromProcessList(db *sqlx.DB, user string, version *version.Version) (string, string, error) {
-	pl := []Processlist{}
-	var err error
-	logs := ""
-	pl, logs, err = GetProcesslistTable(db, version, false, false, false, "50", user)
-	if err != nil {
-		return "N/A", logs, err
-	}
-	for i := range pl {
-		if pl[i].User == user {
-			return strings.Split(pl[i].Host, ":")[0], logs, err
-		}
-	}
-	return "N/A", logs, err
-}
-
-func GetHostFromConnection(db *sqlx.DB, user string, version *version.Version) (string, string, error) {
-	if version == nil {
-		return "N/A", "", errors.New("No database version")
-	}
-	var value string
-	query := "select user()"
-	if version.IsPostgreSQL() {
-		query = "select inet_client_addr()"
-	}
-	err := db.QueryRowx(query).Scan(&value)
-	if err != nil {
-		log.Println("ERROR: Could not get SQL User()", err)
-		return "N/A", query, err
-	}
-	if version.IsPostgreSQL() {
-		return value, query, nil
-	}
-	return strings.Split(value, "@")[1], query, nil
-
-}
-
-func GetPrivileges(db *sqlx.DB, user string, host string, ip string, myver *version.Version) (Privileges, string, error) {
-	db.MapperFunc(strings.Title)
-	stmt := ""
-	var err error
-	priv := Privileges{}
-	if ip == "" {
-		return priv, "", errors.New("Error getting privileges for non-existent IP address")
-	}
-
-	if strings.Contains(ip, ":") {
-		splitip := strings.Split(ip, ":")
-		iprange1 := splitip[0] + ":%:%:%"
-		iprange2 := splitip[0] + ":" + splitip[1] + ":%:%"
-		iprange3 := splitip[0] + ":" + splitip[1] + ":" + splitip[2] + ":%"
-
-		if myver.IsPostgreSQL() {
-			stmt = `SELECT 'Y' as "Select_priv" ,'Y'  as "Process_priv",  CASE WHEN u.usesuper THEN 'Y' ELSE 'N' END  as "Super_priv",  CASE WHEN  u.userepl THEN 'Y' ELSE 'N' END as "Repl_slave_priv", CASE WHEN  u.userepl THEN 'Y' ELSE 'N' END as "Repl_client_priv" ,CASE WHEN u.usesuper THEN 'Y' ELSE 'N' END as "Reload_priv" FROM pg_catalog.pg_user u WHERE u.usename = '` + user + `'`
-			row := db.QueryRowx(stmt)
-			err = row.StructScan(&priv)
-			if err != nil && strings.Contains(err.Error(), "unsupported Scan") {
-				return priv, stmt, errors.New("No replication user defined. Please check the replication user is created with the required privileges")
-			}
-
-		} else {
-			stmt = "SELECT COALESCE(MAX(Select_priv),'N') as Select_priv, COALESCE(MAX(Process_priv),'N') as Process_priv, COALESCE(MAX(Super_priv),'N') as Super_priv, COALESCE(MAX(Repl_slave_priv),'N') as Repl_slave_priv, COALESCE(MAX(Repl_client_priv),'N') as Repl_client_priv, COALESCE(MAX(Reload_priv),'N') as Reload_priv FROM mysql.user WHERE user = ? AND host IN(?,?,?,?,?,?,?,?,?)"
-			row := db.QueryRowx(stmt, user, host, ip, "::", ip+"/255.0.0.0", ip+"/255.255.0.0", ip+"/255:255.255.0", iprange1, iprange2, iprange3)
-			err = row.StructScan(&priv)
-
-			if err != nil && strings.Contains(err.Error(), "unsupported Scan") {
-				return priv, stmt, errors.New("No replication user defined. Please check the replication user is created with the required privileges")
-			}
-		}
-		return priv, stmt, err
-	}
-	splitip := strings.Split(ip, ".")
-
-	iprange1 := splitip[0] + ".%.%.%"
-	iprange4 := splitip[0] + ".%"
-
-	iprange2 := splitip[0] + "." + splitip[1] + ".%.%"
-	iprange5 := splitip[0] + "." + splitip[1] + ".%"
-
-	iprange3 := splitip[0] + "." + splitip[1] + "." + splitip[2] + ".%"
-
-	if myver.IsPostgreSQL() {
-		stmt = `SELECT 'Y' as "Select_priv" ,'Y'  as "Process_priv",  CASE WHEN u.usesuper THEN 'Y' ELSE 'N' END  as "Super_priv",  CASE WHEN  u.userepl THEN 'Y' ELSE 'N' END as "Repl_slave_priv", CASE WHEN  u.userepl THEN 'Y' ELSE 'N' END as "Repl_client_priv" ,CASE WHEN u.usesuper THEN 'Y' ELSE 'N' END as "Reload_priv" FROM pg_catalog.pg_user u WHERE u.usename = '` + user + `'`
-		row := db.QueryRowx(stmt)
-		err = row.StructScan(&priv)
-		if err != nil && strings.Contains(err.Error(), "unsupported Scan") {
-			return priv, stmt, errors.New("No replication user defined. Please check the replication user is created with the required privileges")
-		}
-
-	} else {
-		stmt := "SELECT COALESCE(MAX(Select_priv),'N') as Select_priv, COALESCE(MAX(Process_priv),'N') as Process_priv, COALESCE(MAX(Super_priv),'N') as Super_priv, COALESCE(MAX(Repl_slave_priv),'N') as Repl_slave_priv, COALESCE(MAX(Repl_client_priv),'N') as Repl_client_priv, COALESCE(MAX(Reload_priv),'N') as Reload_priv FROM mysql.user WHERE user = ? AND host IN(?,?,?,?,?,?,?,?,?,?,?)"
-		row := db.QueryRowx(stmt, user, host, ip, "%", ip+"/255.0.0.0", ip+"/255.255.0.0", ip+"/255.255.255.0", iprange1, iprange2, iprange3, iprange4, iprange5)
-		err = row.StructScan(&priv)
-		if err != nil && strings.Contains(err.Error(), "unsupported Scan") {
-			return priv, stmt, errors.New("No replication user defined. Please check the replication user is created with the required privileges")
-		}
-	}
-	return priv, stmt, err
-
-}
-
-func CheckReplicationAccount(db *sqlx.DB, pass string, user string, host string, ip string, myver *version.Version) (bool, string, error) {
-
-	stmt := ""
-	if myver.IsPostgreSQL() {
-		stmt = "SELECT passwd  AS pass ,passwd AS upass  FROM pg_catalog.pg_user u WHERE usename = ?"
-		rows, err := db.Query(stmt, user)
-		if err != nil {
-			return false, stmt, err
-		}
-		for rows.Next() {
-			var pass, upass string
-			err = rows.Scan(&pass, &upass)
-			if err != nil {
-				return false, stmt, err
-			}
-			if pass != upass {
-				return false, stmt, nil
-			}
-		}
-	} else {
-		db.MapperFunc(strings.Title)
-
-		splitip := strings.Split(ip, ".")
-
-		iprange1 := splitip[0] + ".%.%.%"
-		iprange2 := splitip[0] + "." + splitip[1] + ".%.%"
-		iprange3 := splitip[0] + "." + splitip[1] + "." + splitip[2] + ".%"
-
-		stmt = "SELECT STRCMP(Password) AS pass, PASSWORD(?) AS upass FROM mysql.user WHERE user = ? AND host IN(?,?,?,?,?,?,?,?,?)"
-		rows, err := db.Query(stmt, pass, user, host, ip, "%", ip+"/255.0.0.0", ip+"/255.255.0.0", ip+"/255.255.255.0", iprange1, iprange2, iprange3)
-		if err != nil {
-			return false, stmt, err
-		}
-		for rows.Next() {
-			var pass, upass string
-			err = rows.Scan(&pass, &upass)
-			if err != nil {
-				return false, stmt, err
-			}
-			if pass != upass {
-				return false, stmt, nil
-			}
-		}
-	}
-	return true, stmt, nil
-}
-
-func HaveExtraEvents(db *sqlx.DB, file string, pos string) (bool, string, error) {
-	db.MapperFunc(strings.Title)
-	evts := []BinlogEvents{}
-	udb := db.Unsafe()
-	stmt := "SHOW BINLOG EVENTS IN '" + file + "' FROM " + pos
-	err := udb.Get(&evts, stmt)
-	if err != nil {
-		return true, stmt, err
-	}
-	if len(evts) == 1 {
-		return false, stmt, nil
-	}
-	if len(evts) > 1 {
-		return true, stmt, nil
-	}
-	return false, stmt, nil
-}
-
-func GetSlaveStatus(db *sqlx.DB, Channel string, myver *version.Version) (SlaveStatus, string, error) {
-	db.MapperFunc(strings.Title)
-	var err error
-	udb := db.Unsafe()
-	ss := SlaveStatus{}
-	rs := ReplicaStatus{}
-	query := ""
-	if Channel == "" {
-		if myver.IsMySQLOrPerconaGreater84() {
-			query = "SHOW REPLICA STATUS"
-			err = udb.Get(&rs, query)
-		} else {
-			query = "SHOW SLAVE STATUS"
-			if myver.IsPostgreSQL() {
-				/*		query = `select
-							received_lsn ,subname "Connection_name",
-							pg_walfile_name(received_lsn) as "Master_Log_File",
-							(SELECT file_offset  FROM pg_walfile_name_offset(received_lsn)) as "Master_Log_Pos" ,
-							CASE WHEN latest_end_lsn = received_lsn   THEN 0 ELSE EXTRACT(EPOCH FROM latest_end_time -last_msg_send_time) END AS "Seconds_Behind_Master"
-						from pg_catalog.pg_stat_subscription`
-				*/
-				query = `SELECT
-								ss.subname as "Connection_name",
-								ltrim((regexp_split_to_array(s.subconninfo, '\s+'))[2],'host=') as "Master_Host",
-								ltrim((regexp_split_to_array(s.subconninfo, '\s+'))[4],'port=') as "Master_Port",
-								ltrim((regexp_split_to_array(s.subconninfo, '\s+'))[3],'user=') as "Master_User",
-								'master.' || pg_walfile_name(ss.received_lsn) as "Master_Log_File",
-								(SELECT file_offset  FROM pg_walfile_name_offset(ss.received_lsn)) as "Read_Master_Log_Pos" ,
-								'master.' || pg_walfile_name(ss.latest_end_lsn) as "Relay_Master_Log_File",
-								CASE WHEN s.subenabled THEN 'Yes' ELSE 'No' END as "Slave_IO_Running"  ,
-								CASE WHEN s.subenabled THEN 'Yes' ELSE 'No' END as "Slave_SQL_Running",
-									(SELECT file_offset  FROM pg_walfile_name_offset(ss.latest_end_lsn)) as "Exec_Master_Log_Pos",
-								CASE WHEN latest_end_lsn = received_lsn  THEN 0 ELSE EXTRACT(EPOCH FROM latest_end_time -last_msg_send_time) END AS "Seconds_Behind_Master",
-								'' as  "Last_IO_Errno",
-								'' as "Last_SQL_Errno",
-								'' as "Last_SQL_Error" ,
-								0 "Master_Server_Id",
-								'Slave_Pos' as  "Using_Gtid" ,
-								'0-0-' || ('x'|| replace(text(ss.received_lsn), '/' ,''))::bit(64)::bigint  as  "Gtid_IO_Pos" ,
-								'0-0-' || ('x'|| replace(text(ss.latest_end_lsn), '/' ,''))::bit(64)::bigint as "Gtid_Slave_Pos" ,
-								1 as "Slave_Heartbeat_Period" ,
-								'' as "Slave_SQL_Running_State",
-								ros.external_id
-							FROM pg_replication_origin_status ros
-								LEFT JOIN (
-									pg_catalog.pg_stat_subscription ss
-										INNER JOIN  pg_catalog.pg_subscription s
-										ON ss.subname =s.subname
-								) ON ros.external_id='pg_' || ss.subid::text ,
-								(SELECT count(*) as nbrep FROM pg_stat_subscription) AS sqt `
-			}
-
-			err = udb.Get(&ss, query)
-		}
-
-	} else {
-		if myver.IsMariaDB() {
-			query = "SHOW SLAVE '" + Channel + "' STATUS"
-			err = udb.Get(&ss, query)
-		} else if myver.IsMySQLOrPercona() {
-			if myver.GreaterEqual("8.4") {
-				query = "SHOW REPLICA STATUS"
-				err = udb.Get(&rs, query)
-			} else {
-				query = "SHOW SLAVE STATUS FOR CHANNEL '" + Channel + "'"
-				err = udb.Get(&ss, query)
-			}
-		}
-	}
-
-	if myver.IsMySQLOrPerconaGreater84() {
-		ss.ImportFromReplicaStatus(&rs)
-	}
-	//
-	if ss.ChannelName.Valid {
-		if ss.ChannelName.String != "" {
-			ss.ConnectionName.String = ss.ChannelName.String
-			ss.ConnectionName.Valid = true
-		}
-	}
-
-	return ss, query, err
-}
-
-func GetChannelSlaveStatus(db *sqlx.DB, myver *version.Version) ([]SlaveStatus, string, error) {
-	var err error
-	db.MapperFunc(strings.Title)
-	udb := db.Unsafe()
-	rs := []ReplicaStatus{}
-	ss := []SlaveStatus{}
-	uniss := []SlaveStatus{}
-	query := "SHOW SLAVE STATUS"
-	if myver.IsMySQLOrPerconaGreater84() {
-		query = "SHOW REPLICA STATUS"
-		err = udb.Select(&rs, query)
-		if err == nil {
-			for _, r := range rs {
-				s := SlaveStatus{}
-				s.ImportFromReplicaStatus(&r)
-				uniss = append(uniss, s)
-			}
-		}
-	} else {
-		err = udb.Select(&ss, query)
-		// Unified MariaDB MySQL ConnectionName and ChannelName
-		if err == nil {
-			for _, s := range ss {
-				if s.ChannelName.Valid {
-					if s.ChannelName.String != "" {
-						s.ConnectionName.String = s.ChannelName.String
-						s.ConnectionName.Valid = true
-					}
-				}
-				uniss = append(uniss, s)
-			}
-		}
-	}
-
-	return uniss, query, err
-}
-
-func GetPGSlaveStatus(db *sqlx.DB, myver *version.Version) ([]SlaveStatus, error) {
-	db.MapperFunc(strings.Title)
-	udb := db.Unsafe()
-	ss := []SlaveStatus{}
-	query := `SELECT usename, application_name,
-			COALESCE(client_hostname::text, client_addr::text, ''),
-			COALESCE(EXTRACT(EPOCH FROM backend_start)::bigint, 0),
-			backend_xmin, COALESCE(state, ''),
-			COALESCE(sent_lsn::text, ''),
-			COALESCE(write_lsn::text, ''),
-			COALESCE(flush_lsn::text, ''),
-			COALESCE(replay_lsn::text, ''),
-			COALESCE(EXTRACT(EPOCH FROM write_lag)::bigint, 0),
-			COALESCE(EXTRACT(EPOCH FROM flush_lag)::bigint, 0),
-			COALESCE(EXTRACT(EPOCH FROM replay_lag)::bigint, 0),
-			COALESCE(sync_priority, -1),
-			COALESCE(sync_state, ''),
-			pid
-		  FROM pg_stat_replication
-		  ORDER BY pid ASC`
-
-	err := udb.Select(&ss, query)
-
-	return ss, err
-}
-
-func GetDisks(db *sqlx.DB, myver *version.Version) ([]Disk, string, error) {
-	db.MapperFunc(strings.Title)
-	udb := db.Unsafe()
-	ss := []Disk{}
-	query := `SELECT * FROM information_schema.DISKS`
-	err := udb.Select(&ss, query)
-	return ss, query, err
-}
-
-func GetMSlaveStatus(db *sqlx.DB, conn string, myver *version.Version) (SlaveStatus, string, error) {
-
-	s := SlaveStatus{}
-	ss := []SlaveStatus{}
-	var err error
-	logs := ""
-
-	if myver.IsMariaDB() || myver.IsPostgreSQL() {
-		ss, logs, err = GetAllSlavesStatus(db, myver)
-	} else {
-		var s SlaveStatus
-		s, logs, err = GetSlaveStatus(db, conn, myver)
-		ss = append(ss, s)
-	}
-
-	for _, s := range ss {
-		if s.ConnectionName.String == conn {
-			return s, logs, err
-		}
-	}
-	return s, logs, err
-}
-
-func GetAllSlavesStatus(db *sqlx.DB, myver *version.Version) ([]SlaveStatus, string, error) {
-	db.MapperFunc(strings.Title)
-	udb := db.Unsafe()
-	ss := []SlaveStatus{}
-	var err error
-	/*
-
-
-
-
-
-		   type SlaveStatus struct {
-		   	ConnectionName       sql.NullString `db:"Connection_name" json:"connectionName"`
-		   	MasterHost           sql.NullString `db:"Master_Host" json:"masterHost"`
-		   	MasterUser           sql.NullString `db:"Master_User" json:"masterUser"`
-		   	MasterPort           sql.NullString `db:"Master_Port" json:"masterPort"`
-		   	MasterLogFile        sql.NullString `db:"Master_Log_File" json:"masterLogFile"`
-		   	ReadMasterLogPos     sql.NullString `db:"Read_Master_Log_Pos" json:"readMasterLogPos"`
-		   	RelayMasterLogFile   sql.NullString `db:"Relay_Master_Log_File" json:"relayMasterLogFile"`
-		   	SlaveIORunning       sql.NullString `db:"Slave_IO_Running" json:"slaveIoRunning"`
-		   	SlaveSQLRunning      sql.NullString `db:"Slave_SQL_Running" json:"slaveSqlRunning"`
-		   	ExecMasterLogPos     sql.NullString `db:"Exec_Master_Log_Pos" json:"execMasterLogPos"`
-		   	SecondsBehindMaster  sql.NullInt64  `db:"Seconds_Behind_Master" json:"secondsBehindMaster"`
-		   	LastIOErrno          sql.NullString `db:"Last_IO_Errno" json:"lastIoErrno"`
-		   	LastIOError          sql.NullString `db:"Last_IO_Error" json:"lastIoError"`
-		   	LastSQLErrno         sql.NullString `db:"Last_SQL_Errno" json:"lastSqlErrno"`
-		   	LastSQLError         sql.NullString `db:"Last_SQL_Error" json:"lastSqlError"`
-		   	MasterServerID       uint           `db:"Master_Server_Id" json:"masterServerId"`
-		   	UsingGtid            sql.NullString `db:"Using_Gtid" json:"usingGtid"`
-		   	GtidIOPos            sql.NullString `db:"Gtid_IO_Pos" json:"gtidIoPos"`
-		   	GtidSlavePos         sql.NullString `db:"Gtid_Slave_Pos" json:"gtidSlavePos"`
-		   	SlaveHeartbeatPeriod float64        `db:"Slave_Heartbeat_Period" json:"slaveHeartbeatPeriod"`
-		   	ExecutedGtidSet      sql.NullString `db:"Executed_Gtid_Set" json:"executedGtidSet"`
-		   	RetrievedGtidSet     sql.NullString `db:"Retrieved_Gtid_Set" json:"retrievedGtidSet"`
-		   	SlaveSQLRunningState sql.NullString `db:"Slave_SQL_Running_State" json:"slaveSQLRunningState"`
-
-				select * from pg_replication_origin_status
-
-		   }
-	*/
-
-	query := "SHOW ALL SLAVES STATUS"
-	//		CASE WHEN sqt.nbrep=1 THEN	 ss.subname ELSE '' END as "Connection_name",
-	if myver.IsPostgreSQL() {
-		query = `SELECT
-								ss.subname as "Connection_name",
-								ltrim((regexp_split_to_array(s.subconninfo, '\s+'))[2],'host=') as "Master_Host",
-								ltrim((regexp_split_to_array(s.subconninfo, '\s+'))[4],'port=') as "Master_Port",
-								ltrim((regexp_split_to_array(s.subconninfo, '\s+'))[3],'user=') as "Master_User",
-								'master.' || pg_walfile_name(ss.received_lsn) as "Master_Log_File",
-								(SELECT file_offset  FROM pg_walfile_name_offset(ss.received_lsn)) as "Read_Master_Log_Pos" ,
-								'master.' || pg_walfile_name(ss.latest_end_lsn) as "Relay_Master_Log_File",
-								CASE WHEN s.subenabled THEN 'Yes' ELSE 'No' END as "Slave_IO_Running"  ,
-								CASE WHEN s.subenabled THEN 'Yes' ELSE 'No' END as "Slave_SQL_Running",
-									(SELECT file_offset  FROM pg_walfile_name_offset(ss.latest_end_lsn)) as "Exec_Master_Log_Pos",
-								CASE WHEN latest_end_lsn = received_lsn  THEN 0 ELSE EXTRACT(EPOCH FROM latest_end_time -last_msg_send_time) END AS "Seconds_Behind_Master",
-								'' as  "Last_IO_Errno",
-							  '' as "Last_SQL_Errno",
-								'' as "Last_SQL_Error" ,
-								0 "Master_Server_Id",
-							  'Slave_Pos' as  "Using_Gtid" ,
-								'0-0-' || ('x'|| replace(text(ss.received_lsn), '/' ,''))::bit(64)::bigint  as  "Gtid_IO_Pos" ,
-								'0-0-' || ('x'|| replace(text(ss.latest_end_lsn), '/' ,''))::bit(64)::bigint as "Gtid_Slave_Pos" ,
-								1 as "Slave_Heartbeat_Period" ,
-								'' as "Slave_SQL_Running_State",
-								ros.external_id
-							FROM pg_replication_origin_status ros
-							  LEFT JOIN (
-									pg_catalog.pg_stat_subscription ss
-								  	INNER JOIN  pg_catalog.pg_subscription s
-									  ON ss.subname =s.subname
-								) ON ros.external_id='pg_' || ss.subid::text ,
-							  (SELECT count(*) as nbrep FROM pg_stat_subscription) AS sqt `
-	}
-	err = udb.Select(&ss, query)
-	return ss, query, err
-}
-
-func SetMultiSourceRepl(db *sqlx.DB, master_host string, master_port string, master_user string, master_password string, master_filter string) (string, error) {
-	crcTable := crc64.MakeTable(crc64.ECMA) // http://golang.org/pkg/hash/crc64/#pkg-constants
-	checksum64 := fmt.Sprintf("%d", crc64.Checksum([]byte(master_host+":"+master_port), crcTable))
-
-	stmt := "CHANGE MASTER 'mrm_" + checksum64 + "' TO master_host='" + misc.Unbracket(master_host) + "', master_port=" + master_port + ", master_user='" + master_user + "', master_password='" + master_password + "' , master_use_gtid=slave_pos"
-	logs := stmt
-	_, err := db.Exec(stmt)
-	if err != nil {
-		return logs, err
-	}
-	if master_filter != "" {
-
-		stmt = "SET GLOBAL mrm_" + checksum64 + ".replicate_do_table='" + master_filter + "'"
-		logs += "\n" + stmt
-		_, err = db.Exec(stmt)
-		if err != nil {
-			return logs, err
-		}
-	}
-	stmt = "START SLAVE 'mrm_" + checksum64 + "'"
-	logs += "\n" + stmt
-	_, err = db.Exec(stmt)
-	if err != nil {
-		return logs, err
-	}
-
-	return logs, err
-}
-
-func InstallSemiSync(db *sqlx.DB, myver *version.Version) (string, error) {
-	stmt := "INSTALL PLUGIN rpl_semi_sync_slave SONAME 'semisync_slave.so'"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 26)) {
-		stmt = "INSTALL PLUGIN rpl_semi_sync_replica SONAME 'semisync_replica.so'"
-	}
-	logs := stmt
-	_, err := db.Exec(stmt)
-	if err != nil {
-		return logs, err
-	}
-	stmt = "INSTALL PLUGIN rpl_semi_sync_master SONAME 'semisync_master.so'"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 26)) {
-		stmt = "INSTALL PLUGIN rpl_semi_sync_source SONAME 'semisync_source.so';"
-	}
-	logs += "\n" + stmt
-	_, err = db.Exec(stmt)
-	if err != nil {
-		return logs, err
-	}
-	stmt = "set global rpl_semi_sync_master_enabled='ON'"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 26)) {
-		stmt = "SET GLOBAL rpl_semi_sync_source_enabled=ON"
-	}
-	logs += "\n" + stmt
-	_, err = db.Exec(stmt)
-	if err != nil {
-		return logs, err
-	}
-	stmt = "set global rpl_semi_sync_slave_enabled='ON'"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 26)) {
-		stmt = "SET GLOBAL rpl_semi_sync_replica_enabled=ON"
-	}
-	logs += "\n" + stmt
-	_, err = db.Exec(stmt)
-	if err != nil {
-		return logs, err
-	}
-	return logs, nil
-}
-
-func SetBinlogFormat(db *sqlx.DB, format string) (string, error) {
-	query := "set global binlog_format='" + format + "'"
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	return query, nil
-}
-
-func SetBinlogAnnotate(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL binlog_annotate_row_events=ON"
-	logs := query
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	query = "SET GLOBAL replicate_annotate_row_events=ON"
-	logs += "\n" + query
-	_, err = db.Exec(query)
-	if err != nil {
-		return logs, err
-	}
-	return logs, nil
-}
-
-func SetInnoDBLockMonitor(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL innodb_status_output=ON"
-	logs := query
-	_, err := db.Exec(query)
-	if err != nil {
-		return logs, err
-	}
-	query = "SET GLOBAL innodb_status_output_locks=ON"
-	logs += "\n" + query
-	_, err = db.Exec(query)
-	if err != nil {
-		return logs, err
-	}
-	return logs, nil
-}
-
-func UnsetInnoDBLockMonitor(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL innodb_status_output_locks=0"
-	logs := query
-	_, err := db.Exec(query)
-
-	if err != nil {
-		return logs, err
-	}
-	query = "SET GLOBAL innodb_status_output=0"
-	logs += "\n" + query
-	_, err = db.Exec(query)
-	if err != nil {
-		return logs, err
-	}
-	return logs, nil
-}
-
-func SetRelayLogSpaceLimit(db *sqlx.DB, size string) (string, error) {
-	query := "SET GLOBAL relay_log_space_limit=" + size
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	return query, nil
-}
-
-// SetBinlogSlowqueries Enable queries in replication to be reported in slow queries
-func SetBinlogSlowqueries(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL log_slow_slave_statements=ON"
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	return query, nil
-}
-
-func SetLongQueryTime(db *sqlx.DB, querytime string) (string, error) {
-	query := "SET GLOBAL long_query_time=" + querytime
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	return query, nil
-}
-
-// SetSyncBinlog Enable Binlog Durability
-func SetSyncBinlog(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL sync_binlog=1"
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	return query, nil
-}
-
-// SetSyncInnodb Enable InnoDB Durability
-func SetSyncInnodb(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL innodb_flush_log_at_trx_commit=1"
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	return query, nil
-}
-
-// SetBinlogChecksum Enable binlog checksum and check on master
-func SetBinlogChecksum(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL binlog_checksum=1"
-	logs := query
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-
-	query = "SET GLOBAL master_verify_checksum=1"
-	logs += "\n" + query
-	_, err = db.Exec(query)
-	if err != nil {
-		return logs, err
-	}
-	return logs, nil
-}
-
-// SetBinlogCompress Enable MaraiDB 10.2 binlog compression
-func SetBinlogCompress(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL log_bin_compress=1"
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	return query, nil
-}
-
-func SetSlowQueryLogOn(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL slow_query_log=1"
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	return query, nil
-}
-
-func SetSlowQueryLogOff(db *sqlx.DB) (string, error) {
-	query := "SET GLOBAL slow_query_log=0"
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	return query, nil
-}
-
-func ResetAllSlaves(db *sqlx.DB, myver *version.Version) (string, error) {
-
-	ss := []SlaveStatus{}
-	var err error
-	logs := ""
-
-	if myver.IsMariaDB() {
-		ss, logs, err = GetAllSlavesStatus(db, myver)
-	} else {
-		var s SlaveStatus
-		s, logs, err = GetSlaveStatus(db, "", myver)
-		ss = append(ss, s)
-	}
-	if err != nil {
-		return logs, err
-	}
-
-	for _, src := range ss {
-
-		log, err := SetDefaultMasterConn(db, src.ConnectionName.String, myver)
-		logs += "\n" + log
-		if err != nil {
-			return logs, err
-		}
-
-		if myver.IsMySQLOrPercona() {
-			log, _ = StopSlave(db, src.ConnectionName.String, myver)
-			logs += "\n" + log
-		}
-		log, err = ResetSlave(db, true, src.ConnectionName.String, myver)
-		logs += "\n" + log
-		if err != nil {
-			return logs, err
-		}
-	}
-	return logs, err
-}
-
-func GetMasterStatus(db *sqlx.DB, myver *version.Version) (MasterStatus, string, error) {
-	db.MapperFunc(strings.Title)
-	ms := MasterStatus{}
-	udb := db.Unsafe()
-	query := "SHOW MASTER STATUS"
-	if myver.IsPostgreSQL() {
-		query = `select
-		 	 'master.' ||	pg_walfile_name(pg_current_wal_lsn()) as "File" ,
-				(SELECT file_offset  FROM pg_walfile_name_offset(pg_current_wal_lsn())) as "Position" ,
-				'' as Binlog_Do_DB ,
-				'' as "Binlog_Ignore_DB"`
-
-	} else if myver.IsMySQLOrPerconaGreater84() {
-		query = "SHOW BINARY LOG STATUS"
-	}
-	err := udb.Get(&ms, query)
-	//Binlog can be off
-	if err == sql.ErrNoRows {
-		return ms, query, nil
-	}
-	return ms, query, err
-}
-
-func GetSlaveHosts(db *sqlx.DB, myver *version.Version) (map[string]interface{}, string, error) {
-	query := "SHOW SLAVE HOSTS"
-	if myver.IsMySQLOrPerconaGreater84() {
-		query = "SHOW REPLICAS"
-	}
-	rows, err := db.Unsafe().Queryx(query)
-	if err != nil {
-		return nil, query, errors.New("Could not get slave hosts")
-	}
-	defer rows.Close()
-	results := make(map[string]interface{})
-	for rows.Next() {
-		err = rows.MapScan(results)
-		if err != nil {
-			return nil, query, err
-		}
-	}
-	return results, query, nil
-}
-
-func GetSlaveHostsArray(db *sqlx.DB, myver *version.Version) ([]SlaveHosts, string, error) {
-	sh := []SlaveHosts{}
-	query := "SHOW SLAVE HOSTS"
-	if myver.IsMySQLOrPerconaGreater84() {
-		query = "SHOW REPLICAS"
-	}
-	err := db.Unsafe().Select(&sh, query)
-	if err != nil {
-		return nil, query, errors.New("Could not get slave hosts array")
-	}
-	return sh, query, nil
-}
-
-func GetSlaveHostsDiscovery(db *sqlx.DB) ([]string, string, error) {
-	slaveList := []string{}
-	/* This method does not return the server ports, so we cannot rely on it for the time being. */
-	query := "select host from information_schema.processlist where command ='binlog dump'"
-	err := db.Select(&slaveList, query)
-	if err != nil {
-		return nil, query, errors.New("Could not get slave hosts from the processlist")
-	}
-	return slaveList, query, nil
-}
-
-func GetEventStatus(db *sqlx.DB, version *version.Version) ([]Event, string, error) {
-	db.MapperFunc(strings.Title)
-	udb := db.Unsafe()
-
-	ss := []Event{}
-
-	query := "SELECT /*replication-manager*/ db as Db, name as Name, definer as Definer, status+0  AS Status FROM mysql.event"
-	if version.IsMySQLOrPercona() && version.Major >= 8 {
-		query = "SELECT /*replication-manager*/ EVENT_SCHEMA as Db, EVENT_NAME as Name, definer as Definer, status+0  AS Status FROM information_schema.EVENTS"
-	}
-	err := udb.Select(&ss, query)
-	if err != nil {
-		return nil, query, errors.New("Could not get event status")
-	}
-	return ss, query, err
-}
-
-func SetEventStatus(db *sqlx.DB, ev Event, status int64) (string, error) {
-	definer := strings.Split(ev.Definer, "@")
-	if len(definer) != 2 {
-		return "", errors.New("Incorrect definer format")
-	}
-	stmt := fmt.Sprintf("ALTER /*replication-manager*/ DEFINER='%s'@'%s' EVENT ", definer[0], definer[1])
-	if status == 3 {
-		stmt += ev.Db + "." + ev.Name + " DISABLE ON SLAVE"
-	} else {
-		stmt += ev.Db + "." + ev.Name + " ENABLE"
-	}
-	_, err := db.Exec(stmt)
-	if err != nil {
-		return stmt, err
-	}
-	return stmt, nil
-}
-
-func GetVariableSource(db *sqlx.DB, myver *version.Version) string {
-
-	var source string
-	if !myver.IsMariaDB() && ((myver.Major >= 5 && myver.Minor >= 7) || myver.Major >= 6) {
-		source = "performance_schema"
-	} else {
-		source = "information_schema"
-	}
-	return source
-}
-
-func GetStatus(db *sqlx.DB, myver *version.Version, pfs_mutex bool, pfs_latch bool, pfs_mem bool) (map[string]string, string, error) {
-	//fmt.Println("%s: %s", pfs_latch , pfs_mutex,pfs_mem)
-	source := GetVariableSource(db, myver)
-	vars := make(map[string]string)
-	query := "SELECT /*replication-manager*/ UPPER(Variable_name) AS variable_name, UPPER(Variable_Value) AS value FROM " + source + ".global_status"
-	if pfs_mutex {
-		query += " UNION ALL SELECT UPPER(REPLACE(EVENT_NAME,'/','_')) as Variable_name,COUNT_STAR as Value FROM performance_schema.events_waits_summary_global_by_event_name WHERE EVENT_NAME like 'wait/synch/mutex/innodb%' AND COUNT_STAR <>0"
-	}
-	if pfs_latch {
-		query += " UNION ALL SELECT UPPER(REPLACE(EVENT_NAME,'/','_')) as Variable_name,COUNT_STAR as Value FROM performance_schema.events_waits_summary_global_by_event_name WHERE EVENT_NAME like 'wait/synch/rwlock/innodb%' AND COUNT_STAR <>0"
-	}
-
-	if myver.IsPostgreSQL() {
-		query = `SELECT 'COM_QUERY' as "variable_name",  SUM(xact_commit + xact_rollback)::text as "value" FROM pg_stat_database
-			UNION ALL SELECT 'COM_INSERT' as "variable_name",SUM(tup_inserted)::text as "value" FROM pg_stat_database
-			UNION ALL SELECT 'COM_UPDATE' as "variable_name",SUM(tup_updated)::text as "value" FROM pg_stat_database
-			UNION ALL SELECT 'COM_DELETE' as "variable_name",SUM(tup_deleted)::text as "value" FROM pg_stat_database
-			UNION ALL SELECT 'COM_DEADLOCK' as "variable_name",SUM(deadlocks)::text as  "value" FROM pg_stat_database
-			UNION ALL SELECT 'COM_ROLLBACK' as "variable_name",SUM(xact_rollback)::text as  "value" FROM pg_stat_database
-			UNION ALL SELECT 'HANDLER_READ_RND_NEXT' as "variable_name",SUM(tup_fetched)::text as "value" FROM pg_stat_database
-			UNION ALL SELECT 'CREATED_TMP_TABLES' as "variable_name",SUM(temp_files)::text as  "value" FROM pg_stat_database
-			UNION ALL SELECT 'ROWS_SENT' as "variable_name",SUM(tup_returned)::text as  "value" FROM pg_stat_database
-			UNION ALL SELECT 'UPTIME' as "variable_name", EXTRACT(EPOCH FROM pg_postmaster_start_time())::bigint::text  as  "value"
-			UNION ALL SELECT 'THREADS_CONNECTED' as "VARIABLE_NAME",  sum(numbackends)::text  as  "value" FROM pg_stat_database
-			 `
-	}
-	rows, err := db.Queryx(query)
-
-	if err != nil {
-		return nil, query, errors.New("Could not get status variables")
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v Variable
-		err := rows.Scan(&v.Variable_name, &v.Value)
-		if err != nil {
-			return nil, query, errors.New("Could not get results from status scan")
-		}
-		vars[v.Variable_name] = v.Value
-	}
-	if pfs_mem {
-
-		query_pfsmem := "SHOW ENGINE PERFORMANCE_SCHEMA STATUS"
-		rows2, err := db.Queryx(query_pfsmem)
-
-		if err != nil {
-			return nil, query_pfsmem, errors.New("Could not get PFS engine status")
-		}
-		defer rows2.Close()
-		var rtype string
-		var rname string
-		var rvalue uint64
-
-		for rows2.Next() {
-
-			err := rows2.Scan(&rtype, &rname, &rvalue)
-			if err != nil {
-				return nil, query, errors.New("Could not get results from PFS status scan")
-			}
-
-			if strings.Contains(rname, "performance_schema.memory") {
-
-				//	fmt.Println("%s: %s", rname , fmt.Sprint(rvalue))
-
-				vars["performance_schema_memory"] = strconv.FormatUint(rvalue, 10)
-			}
-		}
-	} //end if pfs_mem
-	if vars["ARIA_PAGECACHE_BLOCKS_USED"] != "" {
-		//			fmt.Println(" status ARIA_PAGECACHE_BLOCKS_USED" , fmt.Sprint(vars["ARIA_PAGECACHE_BLOCKS_USED"]))
-		myUInt64, err := strconv.ParseUint(vars["ARIA_PAGECACHE_BLOCKS_USED"], 10, 64)
-		if err == nil {
-			vars["ARIA_PAGECACHE_BYTES_DATA"] = strconv.FormatUint(myUInt64*8192, 10)
-		} else {
-			vars["ARIA_PAGECACHE_BYTES_DATA"] = "0"
-		}
-	} else {
-		vars["ARIA_PAGECACHE_BYTES_DATA"] = "0"
-	}
-	return vars, query, nil
-}
-
-func GetEngineInnoDBStatus(db *sqlx.DB) (string, string, error) {
-	query := "SHOW ENGINE INNODB STATUS"
-	rows, err := db.Query(query)
-	if err != nil {
-		return "", query, err
-	}
-	defer rows.Close()
-	var typeCol, nameCol, statusCol string
-	// First row should contain the necessary info. If many rows returned then it's unknown case.
-	if rows.Next() {
-		if err := rows.Scan(&typeCol, &nameCol, &statusCol); err != nil {
-			return statusCol, query, nil
-		}
-	}
-	return statusCol, query, err
-}
-
-func GetEngineInnoDBVariables(db *sqlx.DB) (map[string]string, string, error) {
-
-	statusCol, logs, err := GetEngineInnoDBStatus(db)
-	if err != nil {
-		return nil, logs, err
-	}
-	vars := make(map[string]string)
-	// 0 queries inside InnoDB, 0 queries in queue
-	// 0 read views open inside InnoDB
-	rQueries, _ := regexp.Compile(`(\d+) queries inside InnoDB, (\d+) queries in queue`)
-	rViews, _ := regexp.Compile(`(\d+) read views open inside InnoDB`)
-	rHistory, _ := regexp.Compile(`History list length (\d+)`)
-	for _, line := range strings.Split(statusCol, "\n") {
-		if data := rQueries.FindStringSubmatch(line); data != nil {
-			vars["queries_inside_innodb"] = data[1]
-			vars["queries_in_queue"] = data[2]
-		} else if data := rViews.FindStringSubmatch(line); data != nil {
-			vars["read_views_open_inside_innodb"] = data[1]
-
-		} else if data := rHistory.FindStringSubmatch(line); data != nil {
-			vars["history_list_lenght_inside_innodb"] = data[1]
-		}
-	}
-	return vars, logs, nil
-}
-
-func EnablePFSQueries(db *sqlx.DB) (string, error) {
-
-	query := "UPDATE setup_consumers SET ENABLED='YES' WHERE NAME IN('events_statements_history_long','events_stages_history')"
-	_, err := db.Exec(query)
-	return query, err
-}
-
-func DisablePFSQueries(db *sqlx.DB) (string, error) {
-
-	query := "UPDATE setup_consumers SET ENABLED='NO' WHERE NAME IN('events_statements_history_long','events_stages_history')"
-	_, err := db.Exec(query)
-	return query, err
-}
-
-func GetSampleQueryFromPFS(db *sqlx.DB, Query PFSQuery) (string, error) {
-	query := "SELECT COALESCE( B.SQL_TEXT,'')  as query FROM performance_schema.events_statements_history_long B WHERE B.DIGEST =''" + Query.Digest + "'"
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return "", err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var res string
-		err := rows.Scan(&res)
-		if err != nil {
-			return "", err
-		}
-		return res, nil
-	}
-	return "", err
-}
-
-func GetQueries(db *sqlx.DB) (map[string]PFSQuery, string, error) {
-
-	vars := make(map[string]PFSQuery)
-	query := "set session group_concat_max_len=2048"
-	db.Exec(query)
-	/*	COALESCE((SELECT B.SQL_TEXT FROM performance_schema.events_statements_history_long B WHERE
-		A.DIGEST = B.DIGEST LIMIT 1 ),'')  as query, */
-	// to expensive FULL SCAN to extact during explain
-	query = `SELECT /*replication-manager*/
-	A.digest as digest,
-	'' as query,
-	A.digest_text as digest_text,
-	A.LAST_SEEN as last_seen,
-	COALESCE(A.SCHEMA_NAME,'') as schema_name,
-	IF(A.SUM_NO_GOOD_INDEX_USED > 0 OR A.SUM_NO_INDEX_USED > 0, '*', '') AS plan_full_scan,
-	A.SUM_CREATED_TMP_DISK_TABLES as plan_tmp_disk,
-	A.SUM_CREATED_TMP_TABLES as plan_tmp_mem,
-	A.COUNT_STAR AS exec_count,
-  A.SUM_ERRORS AS err_count,
-	A.SUM_WARNINGS AS warn_count,
-	SEC_TO_TIME(A.SUM_TIMER_WAIT/1000000000000) AS exec_time_total,
-	(A.MAX_TIMER_WAIT/1000000000000) AS exec_time_max,
-	(A.AVG_TIMER_WAIT/1000000000000) AS exec_time_avg,
-	A.SUM_ROWS_SENT AS rows_sent,
-	ROUND(A.SUM_ROWS_SENT / A.COUNT_STAR) AS rows_sent_avg,
-	A.SUM_ROWS_EXAMINED AS rows_scanned,
-	round(A.sum_timer_wait/1000000000000, 6) as value
-	FROM performance_schema.events_statements_summary_by_digest A
-	WHERE A.digest_text is not null`
-
-	// Do not order as it's eavy fot temporary directory
-	//ORDER BY A.sum_timer_wait desc
-	//LIMIT 50`
-
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return nil, query, errors.New("Could not get queries")
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v PFSQuery
-		err := rows.Scan(&v.Digest, &v.Query, &v.Digest_text, &v.Last_seen, &v.Schema_name, &v.Plan_full_scan, &v.Plan_tmp_disk, &v.Plan_tmp_mem, &v.Exec_count, &v.Err_count, &v.Warn_count, &v.Exec_time_total, &v.Exec_time_max, &v.Exec_time_avg_ms, &v.Rows_sent, &v.Rows_sent_avg, &v.Rows_scanned, &v.Value)
-		if err != nil {
-			return nil, query, errors.New("Could not get results from status scan")
-		}
-		vars[v.Digest] = v
-	}
-	return vars, query, nil
-}
-
-func GetTableChecksumResult(db *sqlx.DB) (map[uint64]chunk, string, error) {
-	vars := make(map[uint64]chunk)
-	query := "SELECT /*replication-manager*/ * from replication_manager_schema.table_checksum"
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return vars, query, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v chunk
-		err = rows.Scan(&v.ChunkId, &v.ChunkMinKey, &v.ChunkMaxKey, &v.ChunkCheckSum)
-		if err != nil {
-			return vars, query, err
-		}
-		vars[v.ChunkId] = v
-	}
-	return vars, query, nil
-}
-
-func GetPlugins(db *sqlx.DB, myver *version.Version) (map[string]*Plugin, string, error) {
-
-	vars := make(map[string]*Plugin)
-	query := `SHOW PLUGINS`
-	if myver.IsMariaDB() {
-		query = `SHOW PLUGINS soname`
-	}
-
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return nil, query, errors.New("Could not get queries")
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v Plugin
-		err := rows.Scan(&v.Name, &v.Status, &v.Type, &v.Library, &v.License)
-		if err != nil {
-			return nil, query, errors.New("Could not get results from plugins scan")
-		}
-		vars[v.Name] = &v
-	}
-	return vars, query, nil
-}
-
-func GetStatusAsInt(db *sqlx.DB, myver *version.Version) (map[string]int64, string, error) {
-	type Variable struct {
-		Variable_name string
-		Value         int64
-	}
-	vars := make(map[string]int64)
-	source := GetVariableSource(db, myver)
-	query := "SELECT /*replication-manager*/ UPPER(Variable_name) AS variable_name, UPPER(Variable_Value) AS value FROM " + source + ".global_status"
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return nil, query, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v Variable
-		rows.Scan(&v.Variable_name, &v.Value)
-		vars[v.Variable_name] = v.Value
-	}
-	return vars, query, nil
-}
-
-func GetVariables(db *sqlx.DB, myver *version.Version) (map[string]string, string, error) {
-	return GetVariablesCase(db, myver, "UPPER")
-}
-
-func GetVariablesCase(db *sqlx.DB, myver *version.Version, vcase string) (map[string]string, string, error) {
-
-	source := GetVariableSource(db, myver)
-	vars := make(map[string]string)
-
-	query := "SELECT /*replication-manager*/ UPPER(Variable_name) AS variable_name, Variable_Value AS value FROM " + source + ".global_variables"
-	if vcase == "UPPER" {
-		query = "SELECT /*replication-manager*/ UPPER(Variable_name) AS variable_name, UPPER(Variable_Value) AS value FROM " + source + ".global_variables"
-	}
-	if myver.IsPostgreSQL() {
-		query = "SELECT upper(name) AS variable_name, setting AS value FROM pg_catalog.pg_settings UNION ALL Select 'SERVER_ID' as variable_name, system_identifier::text as value FROM pg_control_system()"
-		if vcase == "UPPER" {
-			query = "SELECT upper(name) AS variable_name, upper(setting) AS value FROM pg_catalog.pg_settings UNION ALL Select 'SERVER_ID' as variable_name, system_identifier::text as value FROM pg_control_system()"
-		}
-	}
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return vars, query, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v Variable
-		err = rows.Scan(&v.Variable_name, &v.Value)
-		if err != nil {
-			return vars, query, err
-		}
-		vars[v.Variable_name] = v.Value
-	}
-	return vars, query, err
-}
-
-func GetPFSVariablesInstruments(db *sqlx.DB) (map[string]string, string, error) {
-	vars := make(map[string]string)
-	query := "SELECT /*replication-manager*/ UPPER(NAME) AS variable_name, ENABLED AS VALUE from performance_schema.setup_instruments"
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return vars, query, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v Variable
-		err = rows.Scan(&v.Variable_name, &v.Value)
-		if err != nil {
-			return vars, query, err
-		}
-		vars[v.Variable_name] = v.Value
-	}
-	return vars, query, err
-}
-
-func GetPFSVariablesConsumer(db *sqlx.DB) (map[string]string, string, error) {
-
-	vars := make(map[string]string)
-	query := "SELECT /*replication-manager*/ 'SLOW_QUERY_PFS' AS variable_name, IF(count(*)>0,'OFF','ON') AS VALUE from performance_schema.setup_consumers  WHERE NAME IN('events_statements_history_long','events_stages_history') AND ENABLED='NO'"
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return vars, query, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v Variable
-		err = rows.Scan(&v.Variable_name, &v.Value)
-		if err != nil {
-			return vars, query, err
-		}
-		vars[v.Variable_name] = v.Value
-	}
-	return vars, query, err
-}
-
-func GetNoBlockOnMetadataLock(db *sqlx.DB, myver *version.Version) string {
-	if myver.IsPostgreSQL() {
-		return ""
-	}
-	noBlockOnMedataLock := "/*replication-manager*/ "
-	if myver.IsMariaDB() && ((myver.Major == 10 && myver.Minor > 0) || myver.Major > 10) {
-		noBlockOnMedataLock += "SET STATEMENT LOCK_WAIT_TIMEOUT=0 FOR "
-	}
-	return noBlockOnMedataLock
-}
-
-func schemaQuery(db *sqlx.DB, myver *version.Version) string {
-	if myver.IsPostgreSQL() {
-		return `SELECT SCHEMA_NAME AS "SCHEMA_NAME" FROM information_schema.schemata  WHERE SCHEMA_NAME not in ('information_schema','pg_catalog')`
-	}
-
-	return GetNoBlockOnMetadataLock(db, myver) + "SELECT SCHEMA_NAME from information_schema.SCHEMATA WHERE SCHEMA_NAME NOT IN('information_schema','mysql','performance_schema', 'sys') AND SCHEMA_NAME NOT LIKE '#%'"
-}
-
-func getSchemas(db *sqlx.DB, myver *version.Version) ([]string, string, error) {
-	query := schemaQuery(db, myver)
-
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return nil, query, errors.New("could not get schema list")
-	}
-	defer rows.Close()
-
-	var schemas []string
-	for rows.Next() {
-		var s string
-		if err := rows.Scan(&s); err != nil {
-			return nil, query, err
-		}
-		schemas = append(schemas, s)
-	}
-
-	return schemas, query, nil
-}
-
-func tablesQuery(db *sqlx.DB, myver *version.Version, schema string) string {
-	if myver.IsPostgreSQL() {
-		return `SELECT a.schemaname as "Table_schema" ,  a.tablename as "Table_name" ,'postgres' as "Engine",COALESCE(b.n_live_tup,0) as "Table_rows" ,0 as "Data_length",0 as "Index_length" , 0 as "Table_crc"  FROM pg_catalog.pg_tables  a LEFT JOIN pg_catalog.pg_stat_user_tables b ON (a.schemaname=b.schemaname AND a.tablename=b.relname )  WHERE  a.schemaname='` + schema + `'`
-	}
-
-	return GetNoBlockOnMetadataLock(db, myver) + "SELECT a.TABLE_SCHEMA as Table_schema ,  a.TABLE_NAME as Table_name, COALESCE(a.ENGINE,'') as Engine,COALESCE(a.TABLE_ROWS,0) as Table_rows ,COALESCE(a.DATA_LENGTH,0) as Data_length,COALESCE(a.INDEX_LENGTH,0) as Index_length , 0 as Table_crc FROM information_schema.TABLES a WHERE a.TABLE_TYPE='BASE TABLE' AND  a.TABLE_SCHEMA='" + schema + "'"
-}
-
-func ddlQuery(db *sqlx.DB, myver *version.Version, schema, table string) string {
-	if myver.IsPostgreSQL() {
-		return "SELECT 'CREATE TABLE `" + schema + "`.`" + table + "` (' || E'\n'|| '' || string_agg(column_list.column_expr, ', ' || E'\n' || '') ||   '' || E'\n' || ') ENGINE=postgress;' FROM (   SELECT '    `' || column_name || '` ' || data_type ||   coalesce('(' || character_maximum_length || ')', '') ||   case when is_nullable = 'YES' then '' else ' NOT NULL' end as column_expr  FROM information_schema.columns  WHERE table_schema = '" + schema + "' AND table_name = '" + table + "' ORDER BY ordinal_position) column_list"
-	}
-
-	return GetNoBlockOnMetadataLock(db, myver) + "SHOW CREATE TABLE `" + schema + "`.`" + table + "`"
-}
-
-func columnDefQuery(db *sqlx.DB, myver *version.Version, schema string) string {
-	if myver.IsPostgreSQL() {
-		return ""
-	}
-
-	return GetNoBlockOnMetadataLock(db, myver) + "SELECT TABLE_SCHEMA, TABLE_NAME, ORDINAL_POSITION, COLUMN_NAME, COLUMN_TYPE, IS_NULLABLE, COLUMN_DEFAULT, EXTRA, CHARACTER_SET_NAME, COLLATION_NAME FROM information_schema.COLUMNS WHERE TABLE_SCHEMA = '" + schema + "'"
-}
-
-func indexDefQuery(db *sqlx.DB, myver *version.Version, schema string) string {
-	if myver.IsPostgreSQL() {
-		return ""
-	}
-
-	return GetNoBlockOnMetadataLock(db, myver) + "SELECT TABLE_SCHEMA, TABLE_NAME, INDEX_NAME, NON_UNIQUE, INDEX_TYPE, SEQ_IN_INDEX, COLUMN_NAME, SUB_PART FROM information_schema.STATISTICS WHERE TABLE_SCHEMA='" + schema + "'"
-}
-
-func normalizeExtra(e string) string {
-	return strings.ToLower(strings.TrimSpace(e))
-}
-
-func normalizeDefault(d string) *string {
-	d = strings.TrimSpace(d)
-	if strings.EqualFold(d, "NULL") {
-		return nil
-	}
-	return &d
-}
-
-func getTableDDLCRC(
-	db *sqlx.DB,
-	myver *version.Version,
-	schema, table string,
-	crcTable *crc64.Table,
-) (uint64, string) {
-
-	query := ddlQuery(db, myver, schema, table)
-	logs := "\n" + query
-
-	var tbl, ddl string
-	if err := db.QueryRowx(query).Scan(&tbl, &ddl); err != nil {
-		return 0, logs
-	}
-
-	if pos := strings.Index(ddl, "ENGINE="); pos > 0 {
-		ddl = ddl[12:pos]
-	}
-
-	return crc64.Checksum([]byte(ddl), crcTable), logs
-}
-
-type columnRow struct {
-	TableSchema      string
-	TableName        string
-	OrdinalPosition  int
-	ColumnName       string
-	ColumnType       string
-	IsNullable       string
-	ColumnDefault    sql.NullString
-	Extra            string
-	CharacterSetName sql.NullString
-	CollationName    sql.NullString
-}
-
-func getTableColumns(
-	db *sqlx.DB,
-	myver *version.Version,
-	schema string,
-	tables map[string]*Table,
-	crc64Table *crc64.Table,
-) (string, error) {
-
-	query := columnDefQuery(db, myver, schema)
-	logs := "\n" + query
-
-	defer func() {
-		for _, t := range tables {
-			t.HashColumns(crc64Table)
-		}
-	}()
-
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return logs, err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var r columnRow
-		if err := rows.Scan(
-			&r.TableSchema,
-			&r.TableName,
-			&r.OrdinalPosition,
-			&r.ColumnName,
-			&r.ColumnType,
-			&r.IsNullable,
-			&r.ColumnDefault,
-			&r.Extra,
-			&r.CharacterSetName,
-			&r.CollationName,
-		); err != nil {
-			return logs, err
-		}
-
-		key := r.TableSchema + "." + r.TableName
-		t, ok := tables[key]
-		if !ok {
-			t = &Table{
-				TableSchema: r.TableSchema,
-				TableName:   r.TableName,
-			}
-			tables[key] = t
-		}
-
-		col := Column{
-			Name:     r.ColumnName,
-			Type:     strings.ToLower(r.ColumnType),
-			Nullable: r.IsNullable == "YES",
-			Extra:    normalizeExtra(r.Extra),
-		}
-
-		if r.ColumnDefault.Valid {
-			col.Default = normalizeDefault(r.ColumnDefault.String)
-		}
-
-		if r.CharacterSetName.Valid {
-			col.Charset = &r.CharacterSetName.String
-		}
-		if r.CollationName.Valid {
-			col.Collation = &r.CollationName.String
-		}
-
-		t.TableColumns = append(t.TableColumns, col)
-	}
-
-	return logs, nil
-}
-
-type indexRow struct {
-	TableSchema string
-	TableName   string
-	IndexName   string
-	NonUnique   int
-	IndexType   string
-	SeqInIndex  int
-	ColumnName  string
-	SubPart     sql.NullInt64
-}
-
-func getTableIndexes(
-	db *sqlx.DB,
-	myver *version.Version,
-	schema string,
-	tables map[string]*Table,
-	crc64Table *crc64.Table,
-) (string, error) {
-
-	query := indexDefQuery(db, myver, schema)
-	logs := "\n" + query
-
-	defer func() {
-		for _, t := range tables {
-			t.HashIndexes(crc64Table)
-		}
-	}()
-
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return logs, err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var r indexRow
-		if err := rows.Scan(
-			&r.TableSchema,
-			&r.TableName,
-			&r.IndexName,
-			&r.NonUnique,
-			&r.IndexType,
-			&r.SeqInIndex,
-			&r.ColumnName,
-			&r.SubPart,
-		); err != nil {
-			return logs, err
-		}
-
-		key := r.TableSchema + "." + r.TableName
-		t, ok := tables[key]
-		if !ok {
-			// columns loader should have created this,
-			// but be defensive
-			t = &Table{
-				TableSchema: r.TableSchema,
-				TableName:   r.TableName,
-			}
-			tables[key] = t
-		}
-
-		// find or create index
-		var idx *Index
-		for i := range t.TableIndexes {
-			if t.TableIndexes[i].Name == r.IndexName {
-				idx = &t.TableIndexes[i]
-				break
-			}
-		}
-		if idx == nil {
-			idx = &Index{
-				Name:   r.IndexName,
-				Unique: r.NonUnique == 0,
-				Type:   strings.ToUpper(r.IndexType),
-			}
-			t.TableIndexes = append(t.TableIndexes, *idx)
-			idx = &t.TableIndexes[len(t.TableIndexes)-1]
-		}
-
-		// add index column (order preserved by SQL)
-		col := IndexColumn{
-			Name: r.ColumnName,
-		}
-		if r.SubPart.Valid {
-			p := uint16(r.SubPart.Int64)
-			col.Prefix = &p
-		}
-
-		idx.Columns = append(idx.Columns, col)
-	}
-
-	if err := rows.Err(); err != nil {
-		return logs, err
-	}
-
-	return logs, nil
-}
-
-func getTablesForSchema(db *sqlx.DB, myver *version.Version, schema string, crc64Table *crc64.Table) ([]Table, string, error) {
-	var logs string
-	query := tablesQuery(db, myver, schema)
-	logs += "\n" + query
-
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return nil, logs, errors.New("could not get table list: " + err.Error())
-	}
-	defer rows.Close()
-
-	var tables []Table
-
-	for rows.Next() {
-		var t Table
-		if err := rows.Scan(
-			&t.TableSchema,
-			&t.TableName,
-			&t.Engine,
-			&t.TableRows,
-			&t.DataLength,
-			&t.IndexLength,
-			&t.TableCrc,
-		); err != nil {
-			return nil, logs, err
-		}
-
-		crc, qlog := getTableDDLCRC(db, myver, schema, t.TableName, crc64Table)
-		logs += qlog
-		if crc != 0 {
-			t.TableCrc = crc
-		}
-
-		tables = append(tables, t)
-	}
-
-	return tables, logs, nil
-}
-
-func GetTables(db *sqlx.DB, myver *version.Version, getColumns, getIndexes bool) (map[string]*Table, []Table, string, error) {
-
-	var tablemap map[string]*Table = make(map[string]*Table)
-	var tblList []Table
-	logs := ""
-
-	schemas, q, err := getSchemas(db, myver)
-	logs += q
-	if err != nil {
-		return nil, nil, logs, err
-	}
-
-	for _, schema := range schemas {
-		crc64Table := crc64.MakeTable(0xC96C5795D7870F42)
-		var qlog string
-		tables, qlog, err := getTablesForSchema(db, myver, schema, crc64Table)
-		logs += qlog
-		if err != nil {
-			return nil, nil, logs, err
-		}
-
-		for _, t := range tables {
-			key := t.TableSchema + "." + t.TableName
-			tblList = append(tblList, t)
-			tablemap[key] = &t
-		}
-
-		if getColumns {
-			qlog, err := getTableColumns(db, myver, schema, tablemap, crc64Table)
-			logs += qlog
-			if err != nil {
-				return tablemap, tblList, logs, err
-			}
-		}
-
-		if getIndexes {
-			qlog, err := getTableIndexes(db, myver, schema, tablemap, crc64Table)
-			logs += qlog
-			if err != nil {
-				return tablemap, tblList, logs, err
-			}
-		}
-	}
-
-	return tablemap, tblList, logs, nil
-}
-
-func GetUsers(db *sqlx.DB, myver *version.Version) (map[string]*Grant, string, error) {
-	vars := make(map[string]*Grant)
-	// password was remover from system table in mysql 8.0
-
-	query := "SELECT user, host, password, CONV(LEFT(MD5(concat(user,host)), 16), 16, 10)    FROM mysql.user where host<>'localhost' "
-	if myver.IsPostgreSQL() {
-		query = "SELECT usename as user , '%' as host , 'unknow'  as password, 0 FROM pg_catalog.pg_user"
-	} else if myver.IsMySQLOrPercona() && myver.GreaterEqual("5.7.6") {
-		query = "SELECT user, host, authentication_string as password, CONV(LEFT(MD5(concat(user,host)), 16), 16, 10)  FROM mysql.user"
-	}
-
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return nil, query, errors.New("Could not get DB user list")
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var g Grant
-		err = rows.Scan(&g.User, &g.Host, &g.Password, &g.Hash)
-		if err != nil {
-			return vars, query, err
-		}
-		vars["'"+g.User+"'@'"+g.Host+"'"] = &g
-	}
-	return vars, query, nil
-}
-
-func GetProxySQLUsers(db *sqlx.DB) (map[string]Grant, string, error) {
-	vars := make(map[string]Grant)
-	query := "SELECT username, password  FROM mysql_users"
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return nil, query, errors.New("Could not get proxySQL user list")
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var g Grant
-		err = rows.Scan(&g.User, &g.Password)
-		if err != nil {
-			return vars, query, err
-		}
-		vars[g.User+":"+g.Password] = g
-	}
-	return vars, query, nil
-}
-
-func GetSchemas(db *sqlx.DB) ([]string, string, error) {
-	sch := []string{}
-	query := "SELECT SCHEMA_NAME FROM information_schema.SCHEMATA WHERE  SCHEMA_NAME NOT IN('information_schema','mysql','performance_schema','sys') AND SCHEMA_NAME NOT LIKE '#%' "
-	err := db.Select(&sch, query)
-	if err != nil {
-		return nil, query, errors.New("Could not get table list")
-	}
-	return sch, query, nil
-}
-
-func GetSchemasMap(db *sqlx.DB) (map[string]string, string, error) {
-	query := "SELECT SCHEMA_NAME FROM information_schema.SCHEMATA WHERE  SCHEMA_NAME NOT IN('information_schema','mysql','performance_schema')"
-	schemas := make(map[string]string)
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return nil, query, errors.New("Could not get schema list")
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var schema string
-		err = rows.Scan(&schema)
-		if err != nil {
-			return schemas, query, err
-		}
-		schemas[schema] = schema
-	}
-	return schemas, query, nil
-}
-
-func GetVariableByName(db *sqlx.DB, name string, myver *version.Version) (string, string, error) {
-	var value string
-	source := GetVariableSource(db, myver)
-	query := "SELECT Variable_Value AS Value FROM " + source + ".global_variables WHERE Variable_Name = ?"
-	err := db.QueryRowx(query, name).Scan(&value)
-	if err != nil {
-		return "", query, errors.New("Could not get variable by name")
-	}
-	return value, query, nil
-}
-
-func GetVariableByNameToUpper(db *sqlx.DB, name string, myver *version.Version) (string, string, error) {
-	var value string
-	source := GetVariableSource(db, myver)
-	query := "SELECT UPPER(Variable_Value) AS Value FROM " + source + ".global_variables WHERE Variable_Name = ?"
-	err := db.QueryRowx(query, name).Scan(&value)
-	if err != nil {
-		return "", query, errors.New("Could not get variable by name")
-	}
-	return value, query, nil
-}
-
-func FlushBinaryLogsLocal(db *sqlx.DB) (string, error) {
-	_, err := db.Exec("FLUSH LOCAL BINARY LOGS")
-	return "FLUSH LOCAL BINARY LOGS", err
-}
-
-func FlushBinaryLogs(db *sqlx.DB) (string, error) {
-	_, err := db.Exec("FLUSH  BINARY LOGS")
-	return "FLUSH BINARY LOGS", err
-}
-
-func FlushTables(db *sqlx.DB) (string, error) {
-	_, err := db.Exec("FLUSH TABLES")
-	return "FLUSH TABLES", err
-}
-
-func FlushTablesNoLog(db *sqlx.DB) (string, error) {
-	_, err := db.Exec("FLUSH NO_WRITE_TO_BINLOG TABLES")
-	return "FLUSH NO_WRITE_TO_BINLOG TABLES", err
-}
-
-func MariaDBFlushTablesNoLogTimeout(db *sqlx.DB, timeout string) (string, error) {
-	query := "SET STATEMENT max_statement_time=" + timeout + " FOR FLUSH NO_WRITE_TO_BINLOG TABLES"
-	_, err := db.Exec(query)
-	//MySQL does not support DML timeout only SELECT
-	return query, err
-}
-
-func FlushTablesWithReadLock(db *sqlx.DB, myver *version.Version) (string, error) {
-	query := "FLUSH NO_WRITE_TO_BINLOG TABLES WITH READ LOCK"
-	_, err := db.Exec(query)
-	return query, err
-}
-
-func UnlockTables(db *sqlx.DB) (string, error) {
-	query := "UNLOCK TABLES"
-	_, err := db.Exec(query)
-	return query, err
-}
-
-func StopSlave(db *sqlx.DB, Channel string, myver *version.Version) (string, error) {
-	cmd := ""
-	if myver.IsPostgreSQL() {
-		if Channel == "" {
-			Channel = "alltables"
-		}
-		cmd += "ALTER SUBSCRIPTION " + Channel + " DISABLE"
-	} else {
-		if myver.IsMySQLOrPerconaGreater84() {
-			cmd += "STOP REPLICA"
-		} else {
-			cmd += "STOP SLAVE"
-		}
-		if myver.IsMariaDB() && Channel != "" {
-			cmd += " '" + Channel + "'"
-		}
-		if myver.IsMySQLOrPercona() && Channel != "" {
-			cmd += " FOR CHANNEL '" + Channel + "'"
-		}
-	}
-	_, err := db.Exec(cmd)
-	return cmd, err
-}
-
-func StopSlaveIOThread(db *sqlx.DB, Channel string, myver *version.Version) (string, error) {
-	cmd := "STOP SLAVE IO_THREAD"
-	if myver.IsMySQLOrPerconaGreater84() {
-		cmd = "STOP REPLICA IO_THREAD"
-	}
-
-	if myver.IsMariaDB() && Channel != "" {
-		cmd = "STOP SLAVE '" + Channel + "'  IO_THREAD"
-	}
-	if myver.IsMySQLOrPercona() && Channel != "" {
-		cmd += " FOR CHANNEL '" + Channel + "'"
-	}
-	_, err := db.Exec(cmd)
-	return cmd, err
-}
-func StopSlaveSQLThread(db *sqlx.DB, Channel string, myver *version.Version) (string, error) {
-	cmd := "STOP SLAVE SQL_THREAD"
-	if myver.IsMySQLOrPerconaGreater84() {
-		cmd = "STOP REPLICA SQL_THREAD"
-	}
-
-	if myver.IsMariaDB() && Channel != "" {
-		cmd = "STOP SLAVE '" + Channel + "' SQL_THREAD"
-	}
-	if myver.IsMySQLOrPercona() && Channel != "" {
-		cmd += " FOR CHANNEL '" + Channel + "'"
-	}
-	_, err := db.Exec(cmd)
-	return cmd, err
-}
-
-func SetSlaveHeartbeat(db *sqlx.DB, interval string, Channel string, myver *version.Version) (string, error) {
-	var err error
-	logs := ""
-	log := ""
-	log, err = StopSlave(db, Channel, myver)
-	logs += log
-	if err != nil {
-		return logs, err
-	}
-
-	stmt := "CHANGE MASTER TO MASTER_HEARTBEAT_PERIOD=" + interval
-	if myver.IsMySQLOrPerconaGreater84() {
-		stmt = "CHANGE REPLICA SOURCE TO SOURCE_HEARTBEAT_PERIOD=" + interval
-	}
-	logs += "\n" + stmt
-	_, err = db.Exec(stmt)
-
-	if err != nil {
-		return logs, err
-	}
-	log, err = StartSlave(db, Channel, myver)
-	logs += "\n" + stmt
-	if err != nil {
-		return logs, err
-	}
-	return logs, err
-}
-
-func SetSlaveGTIDMode(db *sqlx.DB, mode string, Channel string, myver *version.Version) (string, error) {
-	var err error
-	logs := ""
-	log := ""
-	logs, err = StopSlave(db, Channel, myver)
-	logs += log
-	if err != nil {
-		return logs, err
-	}
-
-	stmt := "CHANGE MASTER TO MASTER_USE_GTID=" + mode
-	if myver.IsMySQLOrPercona() {
-		stmt = "CHANGE MASTER TO MASTER_AUTO_POSITION=1"
-		if myver.GreaterEqual("8.0.23") {
-			stmt = "CHANGE REPLICATION SOURCE TO SOURCE_AUTO_POSITION=1"
-		}
-	}
-
-	logs += "\n" + stmt
-	_, err = db.Exec(stmt)
-	if err != nil {
-		return logs, err
-	}
-
-	log, err = StartSlave(db, Channel, myver)
-	logs += "\n" + stmt
-	if err != nil {
-		return logs, err
-	}
-	return logs, err
-}
-
-func SetSlaveExecMode(db *sqlx.DB, mode string, Channel string, myver *version.Version) (string, error) {
-	var err error
-	logs := ""
-	log := ""
-	logs, err = StopSlave(db, Channel, myver)
-	logs += log
-	if err != nil {
-		return logs, err
-	}
-	stmt := "set global slave_exec_mode='" + mode + "'"
-	logs += "\n" + stmt
-	_, err = db.Exec(stmt)
-	if err != nil {
-		return logs, err
-	}
-	log, err = StartSlave(db, Channel, myver)
-	logs += "\n" + stmt
-	if err != nil {
-		return logs, err
-	}
-	return logs, err
-}
-
-func SetSlaveParallelMode(db *sqlx.DB, mode string, Channel string, myver *version.Version) (string, error) {
-	var err error
-	logs := ""
-	log := ""
-	logs, err = StopSlave(db, Channel, myver)
-	logs += log
-	if err != nil {
-		return logs, err
-	}
-	stmt := "set global slave_parallel_mode='" + mode + "'"
-	logs += "\n" + stmt
-	_, err = db.Exec(stmt)
-	if err != nil {
-		return logs, err
-	}
-	if Channel != "" {
-		stmt := "set global " + Channel + ".slave_parallel_mode='" + mode + "'"
-		_, err = db.Exec(stmt)
-		if err != nil {
-			return logs, err
-		}
-	}
-	log, err = StartSlave(db, Channel, myver)
-	logs += "\n" + stmt
-	if err != nil {
-		return logs, err
-	}
-	return logs, err
-}
-
-func SetGTIDSlavePos(db *sqlx.DB, gtid string) (string, error) {
-	query := "SET GLOBAL gtid_slave_pos='" + gtid + "'"
-	_, err := db.Exec(query)
-	return query, err
-}
-
-func GetBinlogDumpThreads(db *sqlx.DB, myver *version.Version) (int, string, error) {
-	var i int
-	query := "SELECT COUNT(*) AS n FROM INFORMATION_SCHEMA.PROCESSLIST WHERE command LIKE 'binlog dump%'"
-	err := db.Get(&i, query)
-	return i, query, err
-}
-
-func SetMaxConnections(db *sqlx.DB, connections string, myver *version.Version) (string, error) {
-
-	query := "SET GLOBAL max_connections=" + connections
-	_, err := db.Exec(query)
-	return query, err
-}
-
-func SetSemiSyncSlave(db *sqlx.DB, myver *version.Version) (string, error) {
-
-	query := "SET GLOBAL rpl-semi-sync-slave-enabled=1"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 26)) {
-		query = "SET GLOBAL rpl_semi_sync_replica_enabled=ON"
-	}
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	query = "SET GLOBAL rpl-semi-sync-master-enabled=0"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 26)) {
-		query = "SET GLOBAL rpl_semi_sync_source_enabled=OFF"
-	}
-	_, err = db.Exec(query)
-	return query, err
-}
-
-func SetSemiSyncMaster(db *sqlx.DB, myver *version.Version) (string, error) {
-
-	query := "SET GLOBAL rpl-semi-sync-master-enabled=1"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 26)) {
-		query = "SET GLOBAL rpl_semi_sync_source_enabled=ON"
-	}
-	_, err := db.Exec(query)
-	if err != nil {
-		return query, err
-	}
-	query = "SET GLOBAL rpl-semi-sync-slave-enabled=0"
-	if myver.IsMySQLOrPercona() && ((myver.Major >= 8 && myver.Minor > 0) || (myver.Major >= 8 && myver.Minor == 0 && myver.Release >= 26)) {
-		query = "SET GLOBAL rpl_semi_sync_replica_enabled=OFF"
-	}
-	_, err = db.Exec(query)
-	return query, err
-}
-
-func SetSlaveGTIDModeStrict(db *sqlx.DB, myver *version.Version) (string, error) {
-	var err error
-	stmt := ""
-	//MySQL is strict per default with GTID tracking gap trx
-	if myver.IsMariaDB() {
-		stmt = "set global gtid_strict_mode=1"
-		_, err = db.Exec(stmt)
-		if err != nil {
-			return stmt, err
-		}
-	}
-	return stmt, nil
-}
-
-func StopAllSlaves(db *sqlx.DB, myver *version.Version) (string, error) {
-	_, err := db.Exec("STOP ALL SLAVES")
-	return "STOP ALL SLAVES", err
-}
-
-func SkipBinlogEvent(db *sqlx.DB, Channel string, myver *version.Version) (string, error) {
-	if myver.IsMariaDB() {
-		stmt := "SET @@default_master_connection='" + Channel + "'"
-		_, err := db.Exec(stmt)
-		if err != nil {
-			return stmt, err
-		}
-	}
-	query := "SET GLOBAL sql_slave_skip_counter=1"
-	_, err := db.Exec(query)
-	return query, err
-}
-
-func StartSlave(db *sqlx.DB, Channel string, myver *version.Version) (string, error) {
-	cmd := ""
-	if myver.IsPostgreSQL() {
-		if Channel == "" {
-			Channel = "alltables"
-		}
-		cmd += "ALTER SUBSCRIPTION " + Channel + " ENABLE"
-	} else {
-		if myver.IsMySQLOrPerconaGreater84() {
-			cmd += "START REPLICA"
-		} else {
-			cmd += "START SLAVE"
-		}
-
-		if myver.IsMariaDB() && Channel != "" {
-			cmd += " '" + Channel + "'"
-		}
-		if myver.IsMySQLOrPercona() && Channel != "" {
-			cmd += " FOR CHANNEL '" + Channel + "'"
-		}
-	}
-	_, err := db.Exec(cmd)
-	return cmd, err
-}
-
-func StartGroupReplication(db *sqlx.DB, myver *version.Version) (string, error) {
-	cmd := "START GROUP_REPLICATION"
-	_, err := db.Exec(cmd)
-	return cmd, err
-}
-
-func BootstrapGroupReplication(db *sqlx.DB, myver *version.Version) (string, error) {
-	cmd := "SET GLOBAL group_replication_bootstrap_group = ON"
-
-	_, err := db.Exec(cmd)
-	if err != nil {
-		return cmd, err
-	}
-	cmd, err = StartGroupReplication(db, myver)
-	if err != nil {
-		return cmd, err
-	}
-	cmd = "SET GLOBAL group_replication_bootstrap_group = OFF"
-	_, err = db.Exec(cmd)
-
-	return cmd, err
-}
-func ResetSlave(db *sqlx.DB, all bool, Channel string, myver *version.Version) (string, error) {
-	stmt := ""
-	if myver.IsPostgreSQL() {
-		if Channel == "" {
-			Channel = "alltables"
-		}
-		stmt += "DROP SUBSCRIPTION " + Channel
-	} else {
-		if myver.IsMySQLOrPerconaGreater84() {
-			stmt += "RESET REPLICA"
-		} else {
-			stmt += "RESET SLAVE"
-		}
-		if myver.IsMariaDB() && Channel != "" {
-			stmt += " '" + Channel + "'"
-		}
-		if all == true {
-			stmt += " ALL"
-			if myver.IsMySQLOrPercona() && Channel != "" {
-				stmt += " FOR CHANNEL '" + Channel + "'"
-			}
-		}
-	}
-	_, err := db.Exec(stmt)
-	return stmt, err
-}
-
-func ResetMaster(db *sqlx.DB, Channel string, myver *version.Version) (string, error) {
-	stmt := ""
-	if myver.IsPostgreSQL() {
-		if Channel == "" {
-			Channel = "alltables"
-		}
-		stmt += "DROP PUBLICATION " + Channel
-	} else if myver.IsMySQLOrPerconaGreater84() {
-		stmt += "RESET BINARY LOGS AND GTIDS"
-	} else {
-		stmt += "RESET MASTER"
-	}
-	_, err := db.Exec(stmt)
-
-	return stmt, err
-}
-
-func SetGTIDBinlogState(db *sqlx.DB, binlogstate string) (string, error) {
-	stmt := "SET GLOBAL gtid_binlog_state = '" + binlogstate + "'"
-	_, err := db.Exec(stmt)
-
-	return stmt, err
-}
-
-func PostgresGetChannel(db *sqlx.DB, myver *version.Version) (string, string, error) {
-	stmt := ""
-	if myver.IsPostgreSQL() {
-
-		stmt += "select slot_name from pg_replication_slots"
-		channels := []string{}
-		err := db.Select(&channels, stmt)
-		return channels[0], stmt, err
-	}
-	return "", stmt, errors.New("Not PostgreSQL")
-
-}
-
-func SetDefaultMasterConn(db *sqlx.DB, dmc string, myver *version.Version) (string, error) {
-
-	if myver.IsMariaDB() {
-		stmt := "SET @@default_master_connection='" + dmc + "'"
-		_, err := db.Exec(stmt)
-		return stmt, err
-	}
-	// MySQL replication channels are not supported at the moment
-	return "", nil
-}
-
 /*
-	Check for a list of slave prerequisites.
+Package dbhelper provides database helper functions for MariaDB, MySQL, PostgreSQL, and Percona Server.
 
-- Slave is connected
-- Binary log on
-- Connected to master
-- No replication filters
+The package is organized into the following modules:
+
+  - connection.go: Connection management and address resolution
+  - status.go: Server status and variable queries
+  - replication.go: Replication control and monitoring
+  - binlog.go: Binary log operations
+  - transaction.go: Transaction control and locking
+  - performance.go: Performance monitoring and query analysis
+  - tables.go: Table and schema operations
+  - users.go: User and privilege management
+  - events.go: Event scheduler operations
+  - benchmarks.go: Benchmark utilities
+  - groupreplication.go: MySQL Group Replication
+  - spider.go: Spider storage engine operations
+  - types.go: All data type definitions
+
+This refactoring improves code organization, testability, and maintainability.
 */
-func CheckSlavePrerequisites(db *sqlx.DB, s string, myver *version.Version) bool {
-	if debug {
-		log.Printf("CheckSlavePrerequisites called") // remove those warnings !!
-	}
-	err := db.Ping()
-	/* If slave is not online, skip to next iteration */
-	if err != nil {
-		log.Printf("WARN : Slave %s is offline. Skipping", s)
-		return false
-	}
-	vars, _, _ := GetVariables(db, myver)
-	if vars["LOG_BIN"] == "OFF" {
-		return false
-	}
-	return true
-}
-
-func CheckBinlogFilters(m *sqlx.DB, s *sqlx.DB, myver *version.Version) (bool, string, error) {
-	logs := ""
-
-	ms, log, err := GetMasterStatus(m, myver)
-	logs += log
-	if err != nil {
-		return false, log, errors.New("Cannot check binlog status on master")
-	}
-
-	ss, log, err := GetMasterStatus(s, myver)
-	logs += "\n" + log
-	if err != nil {
-		return false, logs, errors.New("ERROR: Can't check binlog status on slave")
-	}
-	if ms.Binlog_Do_DB == ss.Binlog_Do_DB && ms.Binlog_Ignore_DB == ss.Binlog_Ignore_DB {
-		return true, logs, nil
-	}
-	return false, logs, nil
-}
-
-func CheckReplicationFilters(m *sqlx.DB, s *sqlx.DB, myver *version.Version) bool {
-	mv, _, _ := GetVariables(m, myver)
-	sv, _, _ := GetVariables(s, myver)
-	if mv["REPLICATE_DO_TABLE"] == sv["REPLICATE_DO_TABLE"] && mv["REPLICATE_IGNORE_TABLE"] == sv["REPLICATE_IGNORE_TABLE"] && mv["REPLICATE_WILD_DO_TABLE"] == sv["REPLICATE_WILD_DO_TABLE"] && mv["REPLICATE_WILD_IGNORE_TABLE"] == sv["REPLICATE_WILD_IGNORE_TABLE"] && mv["REPLICATE_DO_DB"] == sv["REPLICATE_DO_DB"] && mv["REPLICATE_IGNORE_DB"] == sv["REPLICATE_IGNORE_DB"] {
-		return true
-	} else {
-		return false
-	}
-}
-
-func GetEventScheduler(dbM *sqlx.DB, myver *version.Version) bool {
-
-	sES, _, _ := GetVariableByNameToUpper(dbM, "EVENT_SCHEDULER", myver)
-	if sES != "ON" {
-		return false
-	}
-	return true
-}
-
-func SetEventScheduler(db *sqlx.DB, state bool, myver *version.Version) (string, error) {
-	var err error
-	stmt := ""
-	if state {
-		stmt = "SET GLOBAL event_scheduler=1"
-	} else {
-		stmt = "SET GLOBAL event_scheduler=0"
-	}
-	_, err = db.Exec(stmt)
-	return stmt, err
-}
-
-/* Check if a slave is in sync with his master */
-func CheckSlaveSync(dbS *sqlx.DB, dbM *sqlx.DB, myver *version.Version) bool {
-	if debug {
-		log.Printf("CheckSlaveSync called")
-	}
-	sGtid, _, _ := GetVariableByNameToUpper(dbS, "GTID_CURRENT_POS", myver)
-	mGtid, _, _ := GetVariableByNameToUpper(dbM, "GTID_CURRENT_POS", myver)
-	if sGtid == mGtid {
-		return true
-	} else {
-		return false
-	}
-}
-
-func CheckSlaveSemiSync(dbS *sqlx.DB, myver *version.Version) bool {
-	if debug {
-		log.Printf("CheckSlaveSemiSync called")
-	}
-	sync, _, _ := GetVariableByNameToUpper(dbS, "RPL_SEMI_SYNC_SLAVE_STATUS", myver)
-
-	if sync == "ON" {
-		return true
-	} else {
-		return false
-	}
-}
-
-func MasterWaitGTID(db *sqlx.DB, gtid string, timeout int) (string, error) {
-	query := "SELECT MASTER_GTID_WAIT(?, ?)"
-	_, err := db.Exec(query, gtid, timeout)
-	return query + "(" + gtid + "-" + strconv.Itoa(timeout) + ")", err
-}
-
-func MasterPosWait(db *sqlx.DB, myver *version.Version, log string, pos string, timeout int, channel string) (string, error) {
-	// SOURCE_POS_WAIT  before MySQL 8.0.26
-	funcname := "MASTER_POS_WAIT"
-	if myver.IsMySQLOrPercona() && myver.GreaterEqual("8.0.26") {
-		funcname = "SOURCE_POS_WAIT"
-	}
-
-	if channel == "" {
-		query := "SELECT " + funcname + "(?, ?, ?)"
-		_, err := db.Exec(query, log, pos, timeout)
-		return query + "(" + log + "-" + pos + "-" + strconv.Itoa(timeout) + ")", err
-	} else {
-		query := "SELECT " + funcname + "(?, ?, ?, ?)"
-		_, err := db.Exec(query, log, pos, timeout, channel)
-		return query + "(" + log + "-" + pos + "-" + strconv.Itoa(timeout) + ")", err
-	}
-}
-
-func SetReadOnly(db *sqlx.DB, flag bool) (string, error) {
-	if flag == true {
-		query := "SET GLOBAL read_only=1"
-		_, err := db.Exec(query)
-		return query, err
-	} else {
-		query := "SET GLOBAL read_only=0"
-		_, err := db.Exec(query)
-		return query, err
-	}
-}
-
-func SetSuperReadOnly(db *sqlx.DB, flag bool) (string, error) {
-	if flag == true {
-		_, err := db.Exec("SET GLOBAL super_read_only=1")
-		return "SET GLOBAL super_read_only=1", err
-	} else {
-		_, err := db.Exec("SET GLOBAL super_read_only=0")
-		return "SET GLOBAL super_read_only=0", err
-	}
-}
-
-func SetQueryCaptureMode(db *sqlx.DB, mode string) (string, error) {
-	var err error
-	query := "SET GLOBAL log_output='" + mode + "'"
-
-	if mode == "TABLE" || mode == "FILE" {
-		_, err = db.Exec(query)
-	} else {
-		err = errors.New("Unvalid mode")
-	}
-	return query, err
-}
-
-func CheckLongRunningWrites(db *sqlx.DB, thresh int) (int, string, error) {
-	var count int
-	query := "select SUM(ct) from ( select count(*) as ct from information_schema.processlist  where command = 'Query' and time >= ? and info not like 'select%' union all select count(*) as ct  FROM  INFORMATION_SCHEMA.INNODB_TRX trx WHERE trx.trx_started < CURRENT_TIMESTAMP - INTERVAL ? SECOND) A"
-	err := db.QueryRowx(query, thresh, thresh).Scan(&count)
-	return count, query + "(" + strconv.Itoa(thresh) + ")", err
-}
-
-func KillThreads(db *sqlx.DB, myver *version.Version) (string, error) {
-	//SELECT pg_terminate_backend(11929);
-	var ids []int
-	query := "SELECT Id FROM information_schema.PROCESSLIST WHERE Command != 'binlog dump' AND User != 'system user' AND Id != CONNECTION_ID()"
-	if myver.IsPostgreSQL() {
-		query = "SELECT pid  FROM pg_stat_activity where backend_type='client backend' and pid<>pg_backend_pid()"
-	}
-	logs := query
-	err := db.Select(&ids, query)
-	if err != nil {
-		return logs, err
-	}
-	for _, id := range ids {
-		log, err := KillThread(db, strconv.Itoa(id), myver)
-		logs += log
-		//Should we exit in case of error ?
-		if err != nil {
-			return logs, err
-		}
-	}
-	return logs, err
-
-}
-
-func KillThread(db *sqlx.DB, id string, myver *version.Version) (string, error) {
-	if myver.IsPostgreSQL() {
-		_, err := db.Exec("SELECT pg_terminate_backend(" + id + ")")
-		return "SELECT pg_terminate_backend(" + id + ")", err
-	}
-	_, err := db.Exec("KILL ?", id)
-	return "KILL ? (" + id + ")", err
-}
-
-func KillQuery(db *sqlx.DB, id string, myver *version.Version) (string, error) {
-
-	if myver.IsPostgreSQL() {
-		_, err := db.Exec("SELECT pg_terminate_backend(" + id + ")")
-		return "SELECT pg_terminate_backend(" + id + ")", err
-	}
-	_, err := db.Exec("KILL QUERY ?", id)
-	return "KILL QUERY ? (" + id + ")", err
-
-}
-
-/* Check if string is an IP address or a hostname, return a IP address */
-func CheckHostAddr(h string) (string, error) {
-	var err error
-	if net.ParseIP(h) != nil {
-		return h, err
-	}
-	ha, err := net.LookupHost(h)
-	if err != nil {
-		return "", err
-	}
-	return ha[0], err
-}
-
-func GetSpiderShardUrl(db *sqlx.DB) (string, error) {
-	var value string
-	value = ""
-	err := db.QueryRowx("select coalesce(group_concat(distinct concat(coalesce(st.host,s.host ),':',coalesce(st.port,s.port))),'') as value  from mysql.spider_tables st left join mysql.servers s on st.server=s.server_name").Scan(&value)
-	if err != nil {
-		log.Println("ERROR: Could not get spider shards", err)
-	}
-	return value, err
-}
-
-func GetSpiderMonitor(db *sqlx.DB) (string, error) {
-	var value string
-	value = ""
-	err := db.QueryRowx("select  coalesce(group_concat(distinct concat(coalesce(st.host,s.host ),':',coalesce(st.port,s.port))),'') as value  from mysql.spider_link_mon_servers st left join mysql.servers s on st.server=s.server_name").Scan(&value)
-	if err != nil {
-		log.Println("ERROR: Could not get spider shards", err)
-	}
-	return value, err
-}
-
-func GetSpiderTableToSync(db *sqlx.DB) (map[string]SpiderTableNoSync, error) {
-	vars := make(map[string]SpiderTableNoSync)
-	rows, err := db.Queryx(`
-		select usync.*, sync.srv_sync from (
-		  select  group_concat( distinct concat(db_name, '.',substring_index(table_name,'#P#', 1))) as tbl_src ,  group_concat( distinct concat(db_name, '.', table_name)) as tbl_src_link,concat( coalesce(st.tgt_db_name,s.db) ,'.', tgt_table_name ) as tbl_dest, concat(coalesce(st.host,s.host ),':',coalesce(st.port,s.port)) as srv_desync  from (select * from mysql.spider_tables where link_status=3) st left join mysql.servers s on st.server=s.server_name group by tbl_dest, srv_desync
-		) usync inner join (
-		  select  group_concat( distinct concat(db_name, '.',table_name)) as tbl_src ,concat( coalesce(st.tgt_db_name,s.db) ,'.', tgt_table_name ) as tbl_dest, concat(coalesce(st.host,s.host ),':',coalesce(st.port,s.port)) as srv_sync  from (select * from mysql.spider_tables where link_status=1) st left join mysql.servers s on st.server=s.server_name group by tbl_dest, srv_sync
-		) sync ON  usync.tbl_src_link= sync.tbl_src and usync.tbl_dest=sync.tbl_dest
-		`)
-	if err != nil {
-		return vars, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var v SpiderTableNoSync
-		rows.Scan(&v.Tbl_src, &v.Tbl_src_link, &v.Tbl_dest, &v.Srv_dsync, &v.Srv_sync)
-		vars[v.Tbl_src] = v
-	}
-	return vars, err
-}
-
-func runPreparedExecConcurrent(db *sqlx.DB, n int, co int) error {
-	query := "UPDATE replication_manager_schema.bench SET val=val+1 WHERE id=1"
-	stmt, err := db.Prepare(query)
-	if err != nil {
-		return err
-	}
-
-	remain := int64(n)
-	var wg sync.WaitGroup
-	wg.Add(co)
-
-	for i := 0; i < co; i++ {
-		go func() {
-			for {
-				if atomic.AddInt64(&remain, -1) < 0 {
-					wg.Done()
-					return
-				}
-
-				if _, err1 := stmt.Exec(); err1 != nil {
-					wg.Done()
-					err = err1
-					return
-				}
-			}
-		}()
-	}
-	wg.Wait()
-	stmt.Close()
-	return err
-}
-
-func runPreparedQueryConcurrent(db *sqlx.DB, n int, co int) error {
-	stmt, err := db.Prepare("SELECT ?, \"foobar\"")
-	if err != nil {
-		return err
-	}
-
-	remain := int64(n)
-	var wg sync.WaitGroup
-	wg.Add(co)
-
-	for i := 0; i < co; i++ {
-		go func() {
-			var id int
-			var str string
-			for {
-				if atomic.AddInt64(&remain, -1) < 0 {
-					wg.Done()
-					return
-				}
-
-				if err1 := stmt.QueryRow(i).Scan(&id, &str); err1 != nil {
-					wg.Done()
-					err = err1
-					return
-				}
-			}
-		}()
-	}
-	wg.Wait()
-	stmt.Close()
-	return err
-}
-
-func benchPreparedExecConcurrent1(db *sqlx.DB, n int) error {
-	return runPreparedExecConcurrent(db, n, 1)
-}
-
-func benchPreparedExecConcurrent2(db *sqlx.DB, n int) error {
-	return runPreparedExecConcurrent(db, n, 2)
-}
-
-func benchPreparedExecConcurrent4(db *sqlx.DB, n int) error {
-	return runPreparedExecConcurrent(db, n, 4)
-}
-
-func benchPreparedExecConcurrent8(db *sqlx.DB, n int) error {
-	return runPreparedExecConcurrent(db, n, 8)
-}
-
-func benchPreparedExecConcurrent16(db *sqlx.DB, n int) error {
-	return runPreparedExecConcurrent(db, n, 16)
-}
-
-func InjectLongTrx(db *sqlx.DB, time int) error {
-	benchWarmup(db)
-	_, err := db.Exec("set binlog_format='STATEMENT'")
-	_, err = db.Exec("INSERT INTO replication_manager_schema.bench(val)  select  sleep(" + fmt.Sprintf("%d", time) + ") from dual")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func InjectTrx(db *sqlx.DB) error {
-
-	_, err := db.Exec("INSERT INTO replication_manager_schema.bench(val)  VALUES(1)")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func BenchCleanup(db *sqlx.DB) error {
-	_, err := db.Exec("DROP TABLE replication_manager_schema.bench")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// AnalyzeTable analyzes a table and optionally persists the analysis for all columns or specific columns and indexes.
-// The function returns the executed query and any error encountered.
-func AnalyzeTable(db *sqlx.DB, myver *version.Version, table string, nobinlog, persistent bool, columns string, indexes string) (string, error) {
-	query := "ANALYZE "
-	if nobinlog {
-		query += "LOCAL "
-	}
-	query += "TABLE " + table
-	if myver.Greater("10.4.0") && myver.IsMariaDB() && persistent {
-		if columns == "ALL" {
-			query += " PERSISTENT FOR ALL"
-		} else {
-			query += " PERSISTENT FOR COLUMNS (" + columns + ") INDEXES (" + indexes + ")"
-		}
-	}
-	_, err := db.Exec(query)
-	if err != nil {
-		log.Println("ERROR: Could not analyze table", err)
-	}
-	return query, err
-}
-
-func ChecksumTable(db *sqlx.DB, table string) (string, error) {
-	var tableres string
-	var checkres string
-	tableres = ""
-	checkres = ""
-	err := db.QueryRowx("CHECKSUM TABLE "+table+" EXTENDED").Scan(&tableres, &checkres)
-	if err != nil {
-		log.Println("ERROR: Could not checksum table", err)
-	}
-	return checkres, err
-}
-
-func InjectTrxWithoutCommit(db *sqlx.DB) error {
-	benchWarmup(db)
-	_, err := db.Exec("START TRANSACTION")
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec("INSERT INTO replication_manager_schema.bench(val)  VALUES(1)")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func benchWarmup(db *sqlx.DB) error {
-	db.SetMaxIdleConns(16)
-	_, err := db.Exec("CREATE DATABASE IF NOT EXISTS  replication_manager_schema")
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec("DROP TABLE IF EXISTS replication_manager_schema.bench")
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec("CREATE TABLE IF NOT EXISTS replication_manager_schema.bench(id bigint unsigned primary key auto_increment, val bigint unsigned  )")
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec("INSERT INTO replication_manager_schema.bench(val) VALUES(1)")
-	if err != nil {
-		return err
-	}
-
-	for i := 0; i < 2; i++ {
-		rows, err := db.Query("SELECT val FROM replication_manager_schema.bench")
-		if err != nil {
-			return err
-		}
-
-		if err = rows.Close(); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func WriteConcurrent2(dsn string, qt int) (string, error) {
-	var err error
-
-	bs := BenchmarkSuite{
-		WarmUp:      benchWarmup,
-		Repetitions: 1,
-		PrintStats:  true,
-	}
-
-	if err = bs.AddDriver("mysql", "mysql", dsn); err != nil {
-		return "", err
-	}
-
-	bs.AddBenchmark("PreparedExecConcurrent2", qt, benchPreparedExecConcurrent2)
-
-	result := bs.Run()
-	return result, nil
-}
-
-func IsGroupReplicationMaster(db *sqlx.DB, myver *version.Version, host string) (bool, error) {
-	var value bool
-	value = false
-	err := db.QueryRowx("SELECT 1 FROM  performance_schema.replication_group_members WHERE  MEMBER_STATE='ONLINE' AND MEMBER_ROLE='PRIMARY' AND MEMBER_HOST='" + host + "'").Scan(&value)
-	if err == sql.ErrNoRows {
-		return false, nil
-	}
-	if err != nil {
-		log.Println("ERROR: Could not check Group Replication Master", err)
-	}
-	return value, nil
-}
-
-func IsGroupReplicationSlave(db *sqlx.DB, myver *version.Version, host string) (bool, error) {
-	var value bool
-	value = false
-	err := db.QueryRowx("SELECT 1 FROM  performance_schema.replication_group_members WHERE  MEMBER_STATE='ONLINE' AND MEMBER_ROLE='SECONDARY' AND MEMBER_HOST='" + host + "'").Scan(&value)
-	if err == sql.ErrNoRows {
-		return false, nil
-	}
-	if err != nil {
-		log.Println("ERROR: Could not check Group Replication Secondary", err)
-	}
-	return value, nil
-}
-
-func SetGroupReplicationPrimary(db *sqlx.DB, myver *version.Version) (string, error) {
-	var value string
-	value = ""
-	uuid := ""
-	err := db.QueryRowx("SELECT @@server_uuid").Scan(&uuid)
-	if err != nil {
-		return "", err
-	}
-
-	err = db.QueryRowx("SELECT group_replication_set_as_primary('" + uuid + "')").Scan(&value)
-
-	if err != nil {
-		log.Println("ERROR: Could not set Group Replication Primary", err)
-	}
-	return value, nil
-}
-
-func AddMonitoringUser(db *sqlx.DB, myver *version.Version, user string, password string, method string) error {
-
-	return nil
-}
-
-func AddReplicationUser(db *sqlx.DB, myver *version.Version, user string, password string, method string) error {
-
-	return nil
-}
-
-func CreateUser(db *sqlx.DB, myver *version.Version, user_host string, user_name string, new_password string) (string, error) {
-	query := "CREATE USER '" + user_name + "'@'" + user_host + "' IDENTIFIED BY '" + new_password + "'"
-	_, err := db.Exec(query)
-	if err != nil {
-
-		return query, err
-	}
-	return query, nil
-}
-
-func SetUserPassword(db *sqlx.DB, myver *version.Version, user_host string, user_name string, new_password string) (string, error) {
-	query := "ALTER USER '" + user_name + "'@'" + user_host + "' IDENTIFIED BY '" + new_password + "'"
-	_, err := db.Exec(query)
-	if err != nil {
-
-		return query, err
-	}
-	return query, nil
-}
-
-func RenameUserPassword(db *sqlx.DB, myver *version.Version, user_host string, old_user_name string, new_password string, new_user_name string) (string, error) {
-	query := "RENAME USER '" + old_user_name + "'@'" + user_host + "' TO '" + new_user_name + "'@'" + user_host + "'"
-	_, err := db.Exec(query)
-	if err != nil {
-
-		return query, err
-	}
-	return query, nil
-}
-
-func SetUserGrants(ctx context.Context, conn *sqlx.Conn, myver *version.Version, user_host string, user_name string, grants ...string) (string, error) {
-	var query string
-
-	for _, grant := range grants {
-		query = "GRANT " + grant + " TO '" + user_name + "'@'" + user_host + "'"
-		_, err := conn.ExecContext(ctx, query)
-		if err != nil {
-			return query, err
-		}
-	}
-	return query, nil
-}
-
-func RevokeUserGrants(ctx context.Context, conn *sqlx.Conn, myver *version.Version, user_host string, user_name string) (string, error) {
-	var query string = "REVOKE ALL PRIVILEGES, GRANT OPTION ON *.* FROM '" + user_name + "'@'" + user_host + "'"
-	_, err := conn.ExecContext(ctx, query)
-	return query, err
-}
-
-func SetUserGrantsWithGrantOption(ctx context.Context, conn *sqlx.Conn, myver *version.Version, user_host string, user_name string, grants ...string) (string, error) {
-	var query string
-
-	for _, grant := range grants {
-		query = "GRANT " + grant + " TO '" + user_name + "'@'" + user_host + "' WITH GRANT OPTION"
-		_, err := conn.ExecContext(ctx, query)
-		if err != nil {
-			return query, err
-		}
-	}
-
-	return query, nil
-}
-
-func DuplicateUserPassword(db *sqlx.DB, myver *version.Version, old_user_name string, user_host string, new_user_name string) (string, error) {
-	if myver.IsMySQLOrPercona() && myver.Major >= 8 {
-		query := "SHOW CREATE USER  `" + old_user_name + "`@`" + user_host + "`"
-		rows, err := db.Queryx(query)
-		if err != nil {
-			return query, errors.New("Could not get grant for user ")
-		}
-		defer rows.Close()
-		var grant string
-
-		for rows.Next() {
-			err = rows.Scan(&grant)
-			if err != nil {
-				return query, err
-			}
-			querygrant := strings.Replace(grant, old_user_name, new_user_name, 1)
-			query += ";" + querygrant
-			_, err = db.Queryx(querygrant)
-			if err != nil {
-				return query, err
-			}
-		}
-	}
-	query := "SHOW GRANTS FOR '" + old_user_name + "'@'" + user_host + "'"
-	rows, err := db.Queryx(query)
-	if err != nil {
-		return query, errors.New("Could not get grant for user ")
-	}
-	defer rows.Close()
-	var grant string
-
-	for rows.Next() {
-		err = rows.Scan(&grant)
-		if err != nil {
-			return query, err
-		}
-		querygrant := strings.Replace(grant, old_user_name, new_user_name, 1)
-		query += ";" + querygrant
-		_, err = db.Queryx(querygrant)
-		if err != nil {
-			return query, err
-		}
-	}
-	return query, nil
-}
-
-func PurgeBinlogTo(db *sqlx.DB, filename string) (string, error) {
-	var err error
-	query := "PURGE BINARY LOGS TO '" + filename + "'"
-
-	if filename != "" {
-		_, err = db.Exec(query)
-	} else {
-		return query, errors.New("Invalid filename for PURGE BINARY LOGS TO")
-	}
-	return query, err
-}
-
-func PurgeBinlogBefore(db *sqlx.DB, ts int64) (string, error) {
-	var err error
-	var tstring string = time.Unix(ts, 0).Format(DDMMYYYYhhmmss)
-	query := "PURGE BINARY LOGS BEFORE '" + tstring + "'"
-	_, err = db.Exec(query)
-	return query, err
-}
-
-func SetMaxBinlogTotalSize(db *sqlx.DB, size int) (string, error) {
-	var err error
-	query := "SET GLOBAL max_binlog_total_size = " + strconv.Itoa(size) + ""
-
-	if size >= 0 {
-		_, err = db.Exec(query)
-	} else {
-		return query, errors.New("Invalid size for max_binlog_total_size")
-	}
-	return query, err
-}
-
-func SetSlaveConnectionsNeededForPurge(db *sqlx.DB, size int) (string, error) {
-	var err error
-	query := "SET GLOBAL slave_connections_needed_for_purge = " + strconv.Itoa(size) + ""
-
-	if size >= 0 {
-		_, err = db.Exec(query)
-	} else {
-		return query, errors.New("Invalid value for slave_connections_needed_for_purge")
-	}
-	return query, err
-}
-
-func GetBinlogFormatDesc(db *sqlx.DB, binlogfile string) ([]BinlogEvents, string, error) {
-	logs := ""
-	logpos := "0"
-	events := []BinlogEvents{}
-
-	sql := fmt.Sprintf("show binlog events IN '%s' from %s LIMIT 3", binlogfile, logpos)
-	logs += sql + "\n"
-	err := db.Select(&events, sql)
-	if err != nil {
-		return nil, logs, err
-	}
-
-	for _, row := range events {
-		if strings.ToUpper(row.Event_type) == "FORMAT_DESC" {
-			return []BinlogEvents{row}, logs, nil
-		}
-	}
-
-	return nil, logs, errors.New("Binlog Format Desc Not Found")
-}
-
-func compareDesc(a, b string) int {
-	return map[bool]int{a > b: -1, a < b: 1}[true]
-}
-
-type BinaryLogMetadata struct {
-	Source   string `json:"source"`
-	Filename string `json:"filename"`
-	Start    int64  `json:"start"`
-	Size     uint   `json:"size"`
-}
-
-type BinaryLogMetaMap struct {
-	*sync.Map
-}
-
-func NewBinaryLogMetaMap() *BinaryLogMetaMap {
-	s := new(sync.Map)
-	m := &BinaryLogMetaMap{Map: s}
-	return m
-}
-
-func (m *BinaryLogMetaMap) LoadOrStore(key string, value *BinaryLogMetadata) (*BinaryLogMetadata, bool) {
-	v, ok := m.Map.LoadOrStore(key, value)
-	return v.(*BinaryLogMetadata), ok
-}
-
-func (m *BinaryLogMetaMap) Get(key string) *BinaryLogMetadata {
-	if v, ok := m.Load(key); ok {
-		return v.(*BinaryLogMetadata)
-	}
-	return nil
-}
-
-func (m *BinaryLogMetaMap) CheckAndGet(key string) (*BinaryLogMetadata, bool) {
-	v, ok := m.Load(key)
-	if ok {
-		return v.(*BinaryLogMetadata), true
-	}
-	return nil, false
-}
-
-func (m *BinaryLogMetaMap) Set(key string, value *BinaryLogMetadata) {
-	m.Store(key, value)
-}
-
-func (m *BinaryLogMetaMap) ToNormalMap(c map[string]*BinaryLogMetadata) {
-	// Clear the old values in the output map
-	for k := range c {
-		delete(c, k)
-	}
-
-	// Insert all values from the BinaryLogMetaMap to the output map
-	m.Callback(func(key string, value *BinaryLogMetadata) bool {
-		c[key] = value
-		return true
-	})
-}
-
-func (m *BinaryLogMetaMap) ToNewMap() map[string]BinaryLogMetadata {
-	result := make(map[string]BinaryLogMetadata)
-	m.Range(func(k, v any) bool {
-		result[k.(string)] = *v.(*BinaryLogMetadata)
-		return true
-	})
-	return result
-}
-
-func (m *BinaryLogMetaMap) GetKeys() []string {
-	result := make([]string, 0)
-	m.Range(func(k, v any) bool {
-		result = append(result, k.(string))
-		return true
-	})
-
-	slices.Sort(result)
-	return result
-}
-
-func (m *BinaryLogMetaMap) GetKeysDesc() []string {
-	result := make([]string, 0)
-	m.Range(func(k, v any) bool {
-		result = append(result, k.(string))
-		return true
-	})
-
-	slices.SortFunc(result, compareDesc)
-	return result
-}
-
-func (m *BinaryLogMetaMap) Callback(f func(key string, value *BinaryLogMetadata) bool) {
-	m.Range(func(k, v any) bool {
-		return f(k.(string), v.(*BinaryLogMetadata))
-	})
-}
-
-func (m *BinaryLogMetaMap) Clear() {
-	m.Range(func(key, value any) bool {
-		m.Delete(key.(string))
-		return true
-	})
-}
-
-func FromNormalBinaryLogMetaMap(m *BinaryLogMetaMap, c map[string]*BinaryLogMetadata) *BinaryLogMetaMap {
-	if m == nil {
-		m = NewBinaryLogMetaMap()
-	} else {
-		m.Clear()
-	}
-
-	for k, v := range c {
-		m.Set(k, v)
-	}
-
-	return m
-}
-
-func FromBinaryLogMetaMap(m *BinaryLogMetaMap, c *BinaryLogMetaMap) *BinaryLogMetaMap {
-	if m == nil {
-		m = NewBinaryLogMetaMap()
-	} else {
-		m.Clear()
-	}
-
-	if c != nil {
-		c.Callback(func(key string, value *BinaryLogMetadata) bool {
-			m.Set(key, value)
-			return true
-		})
-	}
-
-	return m
-}
-
-func (m *BinaryLogMetaMap) MarshalIndent(prefix, indent string) ([]byte, error) {
-	result := make(map[string]BinaryLogMetadata, 0)
-	fnames := make([]string, 0)
-	m.Range(func(k, v any) bool {
-		fnames = append(fnames, k.(string))
-		return true
-	})
-
-	slices.Sort(fnames)
-
-	for _, key := range fnames {
-		result[key] = *m.Get(key)
-	}
-
-	return json.MarshalIndent(result, prefix, indent)
-}
-
-func (m *BinaryLogMetaMap) ClearObsoleteMetadata(oldest string) (deleted []string) {
-	keys := make([]string, 0)
-	m.Range(func(k, v any) bool {
-		keys = append(keys, k.(string))
-		return true
-	})
-
-	slices.Sort(keys)
-
-	index := sort.Search(len(keys), func(i int) bool { return keys[i] >= oldest })
-
-	// If the target string is found, return the slice starting from that index
-	if index < len(keys) && keys[index] == oldest {
-		for _, key := range keys[:index] {
-			deleted = append(deleted, key)
-			m.Delete(key)
-		}
-	}
-
-	return deleted
-}
-
-func CountBinaryLogs(db *sqlx.DB, version *version.Version) (int, error) {
-	counter := 0
-	query := "SHOW BINARY LOGS"
-	if version.IsPostgreSQL() {
-		return counter, fmt.Errorf("ERROR: SHOW BINARY LOGS not available on PostgeSQL")
-	}
-	rows, err := db.Queryx(query)
-
-	if err != nil {
-		return counter, errors.New("Could not get binary logs: " + err.Error())
-	}
-	defer rows.Close()
-	for rows.Next() {
-		counter++
-	}
-
-	// Check for any error after iterating through rows
-	if err = rows.Err(); err != nil {
-		return counter, errors.New("Error iterating show binary logs: " + err.Error())
-	}
-
-	return counter, nil
-}
-
-func FetchLogsToBufferTable(conn *sqlx.Conn, version *version.Version, table string, timeStampString string, timeout time.Duration) error {
-	desttablename := table + "_" + timeStampString
-	copytablename := table + "_buffer"
-	query := "USE replication_manager_schema"
-
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to set database as replication_manager_schema: %s", err.Error())
-	}
-	cancel()
-
-	if version.IsMariaDB() {
-		query = "SET SESSION sql_log_bin=0"
-	} else {
-		query = "SET sql_log_bin = OFF"
-	}
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to set sql_log_bin session to 0: %s", err.Error())
-	}
-	cancel()
-
-	query = "DROP TABLE IF EXISTS `replication_manager_schema`." + copytablename
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to create drop replication_manager_schema.%s: %s", copytablename, err.Error())
-	}
-	cancel()
-
-	query = "CREATE TABLE IF NOT EXISTS `replication_manager_schema`." + copytablename + " LIKE `mysql`." + table
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to create table replication_manager_schema.%s: %s", copytablename, err.Error())
-	}
-	cancel()
-
-	query = "CREATE TABLE IF NOT EXISTS `replication_manager_schema`." + desttablename + " LIKE `mysql`." + table
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to create table replication_manager_schema.%s: %s", desttablename, err.Error())
-	}
-	cancel()
-
-	query = "TRUNCATE TABLE `replication_manager_schema`." + copytablename
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to truncate table replication_manager_schema.%s: %s", copytablename, err.Error())
-	}
-	cancel()
-
-	query = "INSERT INTO `replication_manager_schema`." + copytablename + " SELECT * FROM `mysql`." + table
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to create table replication_manager_schema.%s: %s", desttablename, err.Error())
-	}
-	cancel()
-
-	return nil
-}
-
-func TruncateLogsTable(conn *sqlx.Conn, version *version.Version, table string, timeout time.Duration) error {
-	query := "USE mysql"
-
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to set database as replication_manager_schema: %s", err.Error())
-	}
-	cancel()
-
-	if version.IsMariaDB() {
-		query = "SET SESSION sql_log_bin=0"
-	} else {
-		query = "SET sql_log_bin = OFF"
-	}
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to set sql_log_bin session to 0: %s", err.Error())
-	}
-	cancel()
-
-	query = "TRUNCATE TABLE `mysql`." + table
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to truncate table mysql.%s: %s", table, err.Error())
-	}
-	cancel()
-
-	return nil
-}
-
-func MoveLogsToDailyTable(conn *sqlx.Conn, version *version.Version, table string, timeStampString string, timeout time.Duration) error {
-	desttablename := table + "_" + timeStampString
-	copytablename := table + "_buffer"
-
-	query := "USE replication_manager_schema"
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to set database as replication_manager_schema: %s", err.Error())
-	}
-	cancel()
-
-	if version.IsMariaDB() {
-		query = "SET SESSION sql_log_bin=0"
-	} else {
-		query = "SET sql_log_bin = OFF"
-	}
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to set sql_log_bin session to 0: %s", err.Error())
-	}
-	cancel()
-
-	query = "CREATE TABLE IF NOT EXISTS `replication_manager_schema`." + desttablename + " LIKE `mysql`." + table
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to create table replication_manager_schema.%s: %s", desttablename, err.Error())
-	}
-	cancel()
-
-	query = "CREATE TABLE IF NOT EXISTS `replication_manager_schema`." + copytablename + " LIKE `mysql`." + table
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to create table replication_manager_schema.%s: %s", copytablename, err.Error())
-	}
-	cancel()
-
-	query = "INSERT INTO `replication_manager_schema`." + desttablename + " SELECT * FROM `replication_manager_schema`." + copytablename
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to set sql_log_bin session to 0: %s", err.Error())
-	}
-	cancel()
-
-	query = "DROP TABLE IF EXISTS `replication_manager_schema`." + copytablename
-	ctx, cancel = context.WithTimeout(context.Background(), timeout)
-	if _, err := conn.ExecContext(ctx, query); err != nil {
-		cancel()
-		return fmt.Errorf("Unable to set sql_log_bin session to 0: %s", err.Error())
-	}
-	cancel()
-	return nil
-}
-
-func GetBackupSizeEstimation(db *sqlx.DB, version *version.Version) (uint64, error) {
-	var size uint64
-	query := "SELECT SUM(data_length + index_length) AS total_size FROM information_schema.tables"
-	if version.IsPostgreSQL() {
-		return size, fmt.Errorf("ERROR: Backup estimation not available on PostgeSQL")
-	}
-
-	err := db.Get(&size, query)
-	if err != nil {
-		return size, errors.New("Could not get size: " + err.Error())
-	}
-
-	return size, nil
-}
-
-func SetMySQLGtidMode(db *sqlx.DB, mode string) (string, error) {
-	var err error
-	query := "SET GLOBAL gtid_mode = '" + mode + "'"
-
-	if mode == "OFF" || mode == "ON" || mode == "ON_PERMISSIVE" || mode == "OFF_PERMISSIVE" {
-		_, err = db.Exec(query)
-	} else {
-		return query, errors.New("Invalid GTID mode")
-	}
-	return query, err
-}
-
-func SetEnforceGTIDConsistency(db *sqlx.DB, mode string) (string, error) {
-	var err error
-	query := "SET GLOBAL ENFORCE_GTID_CONSISTENCY = '" + mode + "'"
-	if mode == "ON" || mode == "OFF" {
-		_, err = db.Exec(query)
-	} else {
-		return query, errors.New("Invalid GTID mode")
-	}
-	return query, err
-}
+package dbhelper
